Phylota object

Phylota table contains all sequence, cluster and taxonomic information from a phylotaR pipeline run.

# S4 method for Phylota
as.character(x)

# S4 method for Phylota
show(object)

# S4 method for Phylota
print(x)

# S4 method for Phylota
str(object, max.level = 2L, ...)

# S4 method for Phylota
summary(object)

# S4 method for Phylota,character
[[(x, i)

Arguments

x
Phylota object

object
Phylota object

max.level
Maximum level of nesting for str()

...
Further arguments for str()

i
Either sid or cid

Slots

cids

IDs of all clusters

sids

IDs of all sequences

txids

IDs of all taxa

sqs

All sequence records as SeqArc

clstrs

All cluster records as ClstrArc

txdct

Taxonomic dictionary as TaxDict

prnt_id

Parent taxonomic ID

prnt_nm

Parent taxonomic name

See also

Examples

data('aotus')
# this is a Phylota object
# it contains cluster, sequence and taxonomic information from a phylotaR run
show(aotus)

#> Phylota Table (Aotus)
#> - [51] clusters
#> - [335] sequences
#> - [11] source taxa

# you can access its different data slots with @
aotus@cids   # cluster IDs

#>  [1] "0"  "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10" "11" "12" "13" "14"
#> [16] "15" "16" "17" "18" "19" "20" "21" "22" "23" "24" "25" "26" "27" "28" "29"
#> [31] "30" "31" "32" "33" "34" "35" "36" "37" "38" "39" "40" "41" "42" "43" "44"
#> [46] "45" "46" "47" "48" "49" "50"

aotus@sids   # sequence IDs

#>   [1] "AY947712.1/104..396"      "AY947712.1/543..825"     
#>   [3] "AB081277.2/1507..1803"    "AB081271.2/1048..1344"   
#>   [5] "AB081265.2/1048..1344"    "AB181214.1/283..580"     
#>   [7] "AB181211.1/463..760"      "AB181208.1/452..749"     
#>   [9] "AF489256.1/1..298"        "AF489244.1/384..678"     
#>  [11] "AF489244.1/762..1058"     "AY055464.1/137..424"     
#>  [13] "AB181223.1/324..589"      "AB181223.1/719..1043"    
#>  [15] "AB181222.1/324..589"      "AB181222.1/720..1026"    
#>  [17] "AB181222.1/1555..1849"    "AB181221.1/324..589"     
#>  [19] "AB181221.1/719..1042"     "AB181220.1/324..590"     
#>  [21] "AB181220.1/720..1040"     "AB181219.1/324..588"     
#>  [23] "AB181219.1/718..1037"     "AB181218.1/324..588"     
#>  [25] "AB181218.1/718..1038"     "AB181217.1/324..590"     
#>  [27] "AB181217.1/721..1026"     "AB181217.1/1545..1839"   
#>  [29] "AY684992.1/1..1422"       "AY684991.1/1..975"       
#>  [31] "AF201297.1/23..790"       "AF201296.1/23..790"      
#>  [33] "AF201295.1/23..790"       "AF201294.1/23..790"      
#>  [35] "AF201293.1/23..790"       "AF042146.1/11..370"      
#>  [37] "AF042146.1/371..877"      "AF042144.1/11..370"      
#>  [39] "AF042144.1/371..877"      "AY646198.1/152..1576"    
#>  [41] "AF420045.1/255..1619"     "AY055348.1/40..321"      
#>  [43] "NC_018116.1/71..1026"     "NC_018116.1/1095..2647"  
#>  [45] "NC_018116.1/2725..3680"   "NC_018116.1/3892..4930"  
#>  [47] "NC_018116.1/5324..6880"   "NC_018116.1/7009..7696"  
#>  [49] "NC_018116.1/7924..8604"   "NC_018116.1/8604..9387"  
#>  [51] "NC_018116.1/9455..9800"   "NC_018116.1/9868..10164" 
#>  [53] "NC_018116.1/10158..11532" "NC_018116.1/11735..13540"
#>  [55] "NC_018116.1/13537..14070" "NC_018116.1/14143..15282"
#>  [57] "NC_018116.1/15420..16469" "NC_018116.1/15993..16300"
#>  [59] "NC_018115.1/72..1028"     "NC_018115.1/1097..2651"  
#>  [61] "NC_018115.1/2729..3684"   "NC_018115.1/3896..4934"  
#>  [63] "NC_018115.1/5330..6886"   "NC_018115.1/7014..7701"  
#>  [65] "NC_018115.1/7930..8610"   "NC_018115.1/8610..9393"  
#>  [67] "NC_018115.1/9461..9806"   "NC_018115.1/9874..10170" 
#>  [69] "NC_018115.1/10164..11538" "NC_018115.1/11741..13546"
#>  [71] "NC_018115.1/13543..14076" "NC_018115.1/14149..15288"
#>  [73] "NC_018115.1/15426..16585" "NC_018115.1/16115..16417"
#>  [75] "AF368162.1/548..896"      "AF042145.1/1..355"       
#>  [77] "AF042145.1/356..862"      "KC592390.1/71..1027"     
#>  [79] "KC592390.1/1096..2650"    "KC592390.1/2728..3683"   
#>  [81] "KC592390.1/3895..4933"    "KC592390.1/5328..6884"   
#>  [83] "KC592390.1/7012..7699"    "KC592390.1/7928..8608"   
#>  [85] "KC592390.1/8608..9391"    "KC592390.1/9459..9804"   
#>  [87] "KC592390.1/9872..10168"   "KC592390.1/10162..11536" 
#>  [89] "KC592390.1/11739..13544"  "KC592390.1/13541..14074" 
#>  [91] "KC592390.1/14147..15286"  "KC592390.1/15424..16586" 
#>  [93] "KF014091.1/1..1059"       "KF014090.1/1..1059"      
#>  [95] "KF014098.1/1..1017"       "KF014097.1/1..1122"      
#>  [97] "KF014096.1/1..1209"       "KF014095.1/1..1104"      
#>  [99] "KF014099.1/1..1332"       "KF014105.1/1..1113"      
#> [101] "KF014106.1/1..1212"       "KF014108.1/1..927"       
#> [103] "KF014107.1/1..1014"       "KF014110.1/1..1110"      
#> [105] "KF014117.1/1..1263"       "KF014120.1/1..1395"      
#> [107] "KF014122.1/1..978"        "KF014121.1/1..1110"      
#> [109] "KC757385.1/75..1025"      "KC757385.1/1089..2643"   
#> [111] "KC757385.1/2728..3683"    "KC757385.1/3895..4933"   
#> [113] "KC757385.1/5328..6884"    "KC757385.1/7012..7699"   
#> [115] "KC757385.1/7928..8608"    "KC757385.1/8608..9391"   
#> [117] "KC757385.1/9459..9804"    "KC757385.1/9872..10168"  
#> [119] "KC757385.1/10162..11536"  "KC757385.1/11739..13544" 
#> [121] "KC757385.1/13541..14074"  "KC757385.1/14147..15286" 
#> [123] "NC_019799.1/71..1026"     "NC_019799.1/1095..2648"  
#> [125] "NC_019799.1/2726..3680"   "NC_019799.1/3893..4931"  
#> [127] "NC_019799.1/5325..6881"   "NC_019799.1/7009..7682"  
#> [129] "NC_019799.1/7924..8604"   "NC_019799.1/8604..9387"  
#> [131] "NC_019799.1/9455..9800"   "NC_019799.1/9868..10164" 
#> [133] "NC_019799.1/10158..11532" "NC_019799.1/11735..13540"
#> [135] "NC_019799.1/13537..14070" "NC_019799.1/14143..15282"
#> [137] "NC_019799.1/15420..16580" "HQ005473.1/1..1557"      
#> [139] "HQ005473.1/1686..2381"    "HQ005472.1/1..1557"      
#> [141] "HQ005472.1/1686..2381"    "HQ018928.1/177..829"     
#> [143] "HQ018927.1/177..829"      "HQ018926.1/177..829"     
#> [145] "HQ018925.1/177..829"      "HQ018924.1/177..829"     
#> [147] "HQ018923.1/177..829"      "FJ785421.1/71..1026"     
#> [149] "FJ785421.1/1095..2648"    "FJ785421.1/2726..3680"   
#> [151] "FJ785421.1/3893..4931"    "FJ785421.1/5325..6869"   
#> [153] "FJ785421.1/7009..7682"    "FJ785421.1/7924..8604"   
#> [155] "FJ785421.1/8604..9387"    "FJ785421.1/9455..9800"   
#> [157] "FJ785421.1/9868..10164"   "FJ785421.1/10158..11532" 
#> [159] "FJ785421.1/11735..13540"  "FJ785421.1/13537..14070" 
#> [161] "FJ785421.1/14143..15282"  "FJ785421.1/15420..16580" 
#> [163] "AY250707.1/78..1034"      "AY250707.1/1245..2285"   
#> [165] "AY250707.1/2677..4215"    "AY250707.1/4361..5056"   
#> [167] "AY250707.1/5276..5956"    "AY250707.1/5956..6739"   
#> [169] "AY250707.1/6807..7152"    "AY250707.1/7220..7516"   
#> [171] "AY250707.1/7510..8884"    "AY250707.1/9081..10892"  
#> [173] "AY250707.1/10889..11422"  "AY250707.1/11495..12634" 
#> [175] "AF489263.1/353..633"      "AH006732.3/455..733"     
#> [177] "AH006511.2/455..733"      "NC_021939.1/75..1025"    
#> [179] "NC_021939.1/1089..2643"   "NC_021939.1/2728..3683"  
#> [181] "NC_021939.1/3895..4933"   "NC_021939.1/5328..6884"  
#> [183] "NC_021939.1/7012..7699"   "NC_021939.1/7928..8608"  
#> [185] "NC_021939.1/8608..9391"   "NC_021939.1/9459..9804"  
#> [187] "NC_021939.1/9872..10168"  "NC_021939.1/10162..11536"
#> [189] "NC_021939.1/11739..13544" "NC_021939.1/13541..14074"
#> [191] "NC_021939.1/14147..15286" "KR528418.1/1..1140"      
#> [193] "DQ098854.1/1..495"        "DQ098853.1/1..495"       
#> [195] "DQ098852.1/1..495"        "DQ098851.1/1..495"       
#> [197] "HQ005489.1/1..1557"       "HQ005489.1/1685..2380"   
#> [199] "HQ005488.1/1..1557"       "HQ005488.1/1685..2380"   
#> [201] "HQ005487.1/1..1557"       "HQ005487.1/1685..2380"   
#> [203] "HQ005486.1/1..1557"       "HQ005486.1/1685..2380"   
#> [205] "HQ005485.1/1..1557"       "HQ005485.1/1685..2380"   
#> [207] "HQ005484.1/1..1557"       "HQ005484.1/1685..2380"   
#> [209] "HQ005483.1/1..1557"       "HQ005483.1/1685..2380"   
#> [211] "HQ005482.1/1..1557"       "HQ005482.1/1685..2380"   
#> [213] "HQ005481.1/1..1557"       "HQ005481.1/1685..2380"   
#> [215] "HQ005480.1/1..1557"       "HQ005480.1/1685..2380"   
#> [217] "HQ005479.1/1..1557"       "HQ005479.1/1685..2380"   
#> [219] "HQ005478.1/1..1557"       "HQ005478.1/1685..2380"   
#> [221] "HQ005477.1/1..1557"       "HQ005477.1/1685..2380"   
#> [223] "HQ005476.1/1..1557"       "HQ005476.1/1685..2380"   
#> [225] "HQ005475.1/1..1557"       "HQ005475.1/1685..2380"   
#> [227] "HQ005474.1/1..1557"       "HQ005474.1/1685..2380"   
#> [229] "NM_001308516.1/11..370"   "NM_001308516.1/71..367"  
#> [231] "NM_001308516.1/78..356"   "NM_001308516.1/385..877" 
#> [233] "NM_001308533.1/1..1199"   "NM_001308533.1/16..1275" 
#> [235] "NM_001308533.1/385..851"  "NM_001308533.1/409..1102"
#> [237] "NM_001308522.1/61..492"   "NM_001308522.1/184..495" 
#> [239] "NM_001308485.1/482..2098" "NM_001308485.1/530..2095"
#> [241] "NM_001308485.1/530..1531" "KF447738.2/16..285"      
#> [243] "KF447766.1/16..285"       "KF447765.1/16..285"      
#> [245] "KF447764.1/16..285"       "KF447763.1/16..285"      
#> [247] "KF447762.1/16..285"       "KF447761.1/16..285"      
#> [249] "KF447760.1/16..285"       "KF447759.1/16..285"      
#> [251] "KF447758.1/16..285"       "KF447757.1/16..285"      
#> [253] "KF447756.1/16..285"       "KF447755.1/16..285"      
#> [255] "KF447754.1/16..285"       "KF447753.1/16..285"      
#> [257] "KF447752.1/16..285"       "KF447751.1/16..285"      
#> [259] "KF447750.1/16..285"       "KF447749.1/16..285"      
#> [261] "KF447748.1/16..285"       "KF447747.1/16..285"      
#> [263] "KF447746.1/16..285"       "KF447745.1/16..285"      
#> [265] "KF447744.1/16..285"       "KF447743.1/16..285"      
#> [267] "KF447742.1/16..285"       "KF447741.1/16..285"      
#> [269] "KF447740.1/16..285"       "KF447739.1/16..285"      
#> [271] "KF447737.1/16..285"       "KF447736.1/16..285"      
#> [273] "KF447735.1/16..285"       "KF447734.1/16..285"      
#> [275] "KF447733.1/16..285"       "NM_001308528.1/23..790"  
#> [277] "AF397604.1/737..1037"     "AF338376.1/177..829"     
#> [279] "AF338375.1/177..829"      "AF338374.1/177..829"     
#> [281] "AF352260.1/34..729"       "AF352259.1/34..729"      
#> [283] "AF352258.1/34..729"       "AF352257.1/34..729"      
#> [285] "AF352256.1/34..729"       "AF352255.1/34..729"      
#> [287] "AF352254.1/34..729"       "JN161101.1/71..1026"     
#> [289] "JN161101.1/1095..2648"    "JN161101.1/2726..3681"   
#> [291] "JN161101.1/3893..4931"    "JN161101.1/5325..6881"   
#> [293] "JN161101.1/7010..7697"    "JN161101.1/7925..8605"   
#> [295] "JN161101.1/8605..9388"    "JN161101.1/9456..9801"   
#> [297] "JN161101.1/9869..10165"   "JN161101.1/10159..11533" 
#> [299] "JN161101.1/11736..13541"  "JN161101.1/13538..14071" 
#> [301] "JN161101.1/14144..15283"  "JN161101.1/15421..16471" 
#> [303] "JN161101.1/15994..16302"  "JN161100.1/71..1026"     
#> [305] "JN161100.1/1095..2647"    "JN161100.1/2725..3680"   
#> [307] "JN161100.1/3892..4930"    "JN161100.1/5324..6880"   
#> [309] "JN161100.1/7009..7696"    "JN161100.1/7924..8604"   
#> [311] "JN161100.1/8604..9387"    "JN161100.1/9455..9800"   
#> [313] "JN161100.1/9868..10164"   "JN161100.1/10158..11532" 
#> [315] "JN161100.1/11735..13540"  "JN161100.1/13537..14070" 
#> [317] "JN161100.1/14143..15282"  "JN161100.1/15420..16469" 
#> [319] "JN161100.1/15993..16300"  "JN161099.1/72..1028"     
#> [321] "JN161099.1/1097..2651"    "JN161099.1/2729..3684"   
#> [323] "JN161099.1/3896..4934"    "JN161099.1/5330..6886"   
#> [325] "JN161099.1/7014..7701"    "JN161099.1/7930..8610"   
#> [327] "JN161099.1/8610..9393"    "JN161099.1/9461..9806"   
#> [329] "JN161099.1/9874..10170"   "JN161099.1/10164..11538" 
#> [331] "JN161099.1/11741..13546"  "JN161099.1/13543..14076" 
#> [333] "JN161099.1/14149..15288"  "JN161099.1/15426..16585" 
#> [335] "JN161099.1/16115..16417" 

aotus@txids  # taxonomic IDs

#>  [1] "30591"  "280755" "292213" "9505"   "37293"  "120088" "867331" "57176" 
#>  [9] "43147"  "57175"  "231953"

aotus@clstrs # clusters archive

#> Archive of cluster record(s)
#>  - [51] clusters

aotus@sqs    # sequence archive

#> Archive of sequence record(s)
#>  - [335] sequences
#>  - [11] unique txids
#>  - [696] median sequence length
#>  - [0] median ambiguous nucleotides

aotus@txdct  # taxonomic dictionary

#> Taxonomic dictionary [21] recs, parent [id 9504]

# see all of the available slots
(slotNames(aotus))

#> [1] "cids"    "txids"   "sids"    "txdct"   "sqs"     "clstrs"  "prnt_id"
#> [8] "prnt_nm"

# access different sequences and clusters with [[
(aotus[['0']])              # cluster record 0

#> Cluster Record [id 0]
#>  - [subtree] type
#>  - [AF352254.1/34..729] seed sequence
#>  - [36] sequences
#>  - [10] taxa

(aotus[[aotus@sids[[1]]]])  # first sequence record

#> SeqRec [ID: AY947712.1/104..396]

# get a summary of the whole object
(summary(aotus))

#>    ID    Type                     Seed Parent N_taxa N_seqs Med_sql       MAD
#> 1   0 subtree       AF352254.1/34..729   9504     10     36   696.0 0.9956897
#> 2   1 subtree       KF447735.1/16..285   9504      2     34   270.0 1.0000000
#> 3   2 subtree    AY250707.1/2677..4215   9504     10     29  1557.0 0.9993356
#> 4   3 subtree       KF014117.1/1..1263   9504      1     16  1110.0 0.8075269
#> 5   4 subtree     AB181218.1/718..1038   9504      4     14   306.5 0.8794515
#> 6   5 subtree  AY250707.1/11495..12634   9504      7     12  1140.0 1.0000000
#> 7   6 subtree     AF420045.1/255..1619   9504      5     11  1554.0 0.9884829
#> 8   7 subtree      AY250707.1/78..1034   9504      6     11   956.0 0.9988601
#> 9   8 subtree    AY250707.1/1245..2285   9504      6     11  1039.0 0.9982534
#> 10  9 subtree    AY250707.1/5276..5956   9504      6     11   681.0 1.0000000
#> 11 10 subtree    AY250707.1/5956..6739   9504      6     11   784.0 1.0000000
#> 12 11 subtree    AY250707.1/6807..7152   9504      6     11   346.0 1.0000000
#> 13 12 subtree    AY250707.1/7220..7516   9504      6     11   297.0 1.0000000
#> 14 13 subtree    AY250707.1/7510..8884   9504      6     11  1375.0 1.0000000
#> 15 14 subtree   AY250707.1/9081..10892   9504      6     11  1806.0 0.9969898
#> 16 15 subtree  AY250707.1/10889..11422   9504      6     11   534.0 1.0000000
#> 17 16 subtree      FJ785421.1/71..1026   9504      5     10   956.0 0.9982236
#> 18 17 subtree      AF338374.1/177..829   9504      8      9   653.0 1.0000000
#> 19 18 subtree       AF042144.1/11..370   9504      3      8   326.0 0.8920139
#> 20 19 subtree  FJ785421.1/15420..16580   9504      4      8  1160.0 0.9625967
#> 21 20 subtree       HQ005472.1/1..1557  30591      3      8  1557.0 1.0000000
#> 22 21 subtree    HQ005472.1/1686..2381  30591      3      8   696.0 0.9956897
#> 23 22 subtree      AB181217.1/324..590   9504      2      7   266.0 0.9962547
#> 24 23 subtree    AB081265.2/1048..1344   9504      2      6   297.5 0.9983221
#> 25 24 subtree       AF201293.1/23..790   9504      1      6   768.0 1.0000000
#> 26 25 subtree  JN161099.1/16115..16417   9504      2      5   308.0 0.9909385
#> 27 26 subtree      AB181217.1/324..590  30591      1      5   266.0 0.9962547
#> 28 27 subtree     AB181217.1/721..1026  30591      1      5   321.0 0.9827160
#> 29 28 subtree    AB181217.1/1545..1839   9504      3      4   295.0 0.9924497
#> 30 29 subtree      AF042144.1/371..877   9504      2      4   507.0 0.9930966
#> 31 30 subtree NM_001308533.1/409..1102   9504      1      4   946.5 0.7182540
#> 32 31  direct        DQ098851.1/1..495  30591      1      4   495.0 1.0000000
#> 33 32 subtree    AB081277.2/1507..1803  30591      1      4   298.0 0.9991611
#> 34 33 subtree     AY646198.1/152..1576   9504      1      3  1422.0 0.8940351
#> 35 34 subtree     AF397604.1/737..1037   9504      2      3   312.0 0.8063272
#> 36 35 subtree NM_001308485.1/482..2098   9504      1      3  1566.0 0.8627087
#> 37 36  direct     AF420045.1/255..1619  30591      1      3  1555.0 0.9592712
#> 38 37 subtree      JN161099.1/72..1028  30591      2      3   957.0 1.0000000
#> 39 38 subtree    JN161099.1/1097..2651  30591      2      3  1555.0 1.0000000
#> 40 39 subtree    JN161099.1/2729..3684  30591      2      3   956.0 1.0000000
#> 41 40 subtree    JN161099.1/3896..4934  30591      2      3  1039.0 1.0000000
#> 42 41 subtree    JN161099.1/7930..8610  30591      2      3   681.0 1.0000000
#> 43 42 subtree    JN161099.1/8610..9393  30591      2      3   784.0 1.0000000
#> 44 43 subtree    JN161099.1/9461..9806  30591      2      3   346.0 1.0000000
#> 45 44 subtree   JN161099.1/9874..10170  30591      2      3   297.0 1.0000000
#> 46 45 subtree  JN161099.1/10164..11538  30591      2      3  1375.0 1.0000000
#> 47 46 subtree  JN161099.1/11741..13546  30591      2      3  1806.0 1.0000000
#> 48 47 subtree  JN161099.1/13543..14076  30591      2      3   534.0 1.0000000
#> 49 48 subtree  JN161099.1/14149..15288  30591      2      3  1140.0 1.0000000
#> 50 49 subtree  JN161099.1/15426..16585  30591      2      3  1160.0 0.9982803
#> 51 50 subtree      AF338376.1/177..829  30591      2      3   653.0 1.0000000
#>                         Definition                            Feature
#> 1      complete (0.1), gene (0.08)              co2 (0.5), cox2 (0.3)
#> 2    allele (0.08), antigen (0.08)       aovodrb (0.5), aonadrb (0.5)
#> 3       complete (0.1), and (0.07)              co1 (0.6), cox1 (0.3)
#> 4      allele (0.08), aotus (0.08)       kir3ds3 (0.2), kir2ds5 (0.1)
#> 5       aotus (0.1), azarai (0.07)            alu (0.1), alusg (0.06)
#> 6      aotus (0.2), complete (0.2)                           cytb (1)
#> 7      aotus (0.1), complete (0.1)         16s (0.3), ribosomal (0.3)
#> 8        aotus (0.2), genome (0.2)                            nd1 (1)
#> 9        aotus (0.2), genome (0.2)                            nd2 (1)
#> 10       aotus (0.2), genome (0.2)          atp6 (0.9), atpase (0.09)
#> 11       aotus (0.2), genome (0.2)           cox3 (0.9), coiii (0.09)
#> 12       aotus (0.2), genome (0.2)                            nd3 (1)
#> 13       aotus (0.2), genome (0.2)                           nd4l (1)
#> 14       aotus (0.2), genome (0.2)                            nd4 (1)
#> 15       aotus (0.2), genome (0.2)                            nd5 (1)
#> 16       aotus (0.2), genome (0.2)                            nd6 (1)
#> 17     aotus (0.2), complete (0.2)         12s (0.3), ribosomal (0.3)
#> 18          sry (0.2), aotus (0.1)                            sry (1)
#> 19         aotus (0.1), mrna (0.1)          aotusb2m (0.5), b2m (0.5)
#> 20     aotus (0.2), complete (0.2)        control (0.4), region (0.4)
#> 21      complete (0.1), and (0.07)              co1 (0.6), cox1 (0.4)
#> 22      complete (0.1), and (0.07)              co2 (0.6), cox2 (0.4)
#> 23 aotus (0.07), downstream (0.07)       mlws14x (0.1), mlws14y (0.1)
#> 24       aotus (0.08), cone (0.08)   lwsmws14x (0.2), lwsmws14y (0.2)
#> 25    antigen (0.08), aotus (0.08) aonadqa1 (0.8), loc105708941 (0.2)
#> 26     aotus (0.2), complete (0.2)        hstrand (0.5), origin (0.5)
#> 27     aotus (0.07), azarai (0.07)      mlws29x (0.2), mlws29y1 (0.2)
#> 28     aotus (0.07), azarai (0.07)      mlws29x (0.2), mlws29y1 (0.2)
#> 29         aotus (0.1), gene (0.1)         sinealu (0.3), color (0.1)
#> 30         aotus (0.1), mrna (0.1)          aotusb2m (0.8), b2m (0.2)
#> 31        aotus (0.1), group (0.1)                          nr3c1 (1)
#> 32       aotus (0.1), azarai (0.1)                        trimcyp (1)
#> 33     aotus (0.08), azarai (0.08)    lwsmws29x (0.2), mlws29y1 (0.2)
#> 34          aotus (0.1), cds (0.1)          of (0.1), activity (0.07)
#> 35         aotus (0.1), beta (0.1)      loc105722736 (0.7), cgb (0.3)
#> 36   aotus (0.1), endogenous (0.1)                        ervfrd1 (1)
#> 37       aotus (0.1), azarai (0.1)         16s (0.3), ribosomal (0.3)
#> 38       azarai (0.3), aotus (0.2)         12s (0.3), ribosomal (0.3)
#> 39       azarai (0.3), aotus (0.2)         16s (0.3), ribosomal (0.3)
#> 40       azarai (0.3), aotus (0.2)                            nd1 (1)
#> 41       azarai (0.3), aotus (0.2)                            nd2 (1)
#> 42       azarai (0.3), aotus (0.2)                           atp6 (1)
#> 43       azarai (0.3), aotus (0.2)                           cox3 (1)
#> 44       azarai (0.3), aotus (0.2)                            nd3 (1)
#> 45       azarai (0.3), aotus (0.2)                           nd4l (1)
#> 46       azarai (0.3), aotus (0.2)                            nd4 (1)
#> 47       azarai (0.3), aotus (0.2)                            nd5 (1)
#> 48       azarai (0.3), aotus (0.2)                            nd6 (1)
#> 49       azarai (0.3), aotus (0.2)                           cytb (1)
#> 50       azarai (0.3), aotus (0.2)        control (0.4), region (0.4)
#> 51          sry (0.2), aotus (0.1)                            sry (1)

# the above generates a data.frame with information on each cluster:
# ID - unique id in the object
# Type - cluster type
# Seed - most connected sequence
# Parent - MRCA of all represented taxa
# N_taxa - number of NCBI recognised taxa
# N_seqs - number of sequences
# Med_sql - median sequence length
# MAD - Maximum alignment density, values close to 1 indicate all sequences in
#  the cluster have a similar length.
# Definition - most common words (and frequency) in sequence definitions
# Feature - most common feature name (and frequency)