Sequence record

Sequence record contains sequence data.

# S4 method for SeqRec
as.character(x)

# S4 method for SeqRec
show(object)

# S4 method for SeqRec
print(x)

# S4 method for SeqRec
str(object, max.level = 2L, ...)

# S4 method for SeqRec
summary(object)

Arguments

x
SeqRec object

object
SeqRec object

max.level
Maximum level of nesting for str()

...
Further arguments for str()

Details

Sequence is stored as raw. Use rawToChar().

Slots

id

Unique ID

nm

Best-guess sequence name

accssn

Accession

vrsn

Accession version

url

URL

txid

Taxonomic ID of source taxon

orgnsm

Scientific name of source taxon

sq

Sequence

dfln

Definition line

ml_typ

Molecule type, e.g. DNA

rec_typ

Record type: Whole or feature

nncltds

Number of nucleotides

nambgs

Number of ambiguous nucleotides

pambgs

Proportion of ambiguous nucleotides

gcr

GC ratio

age

Number of days between sequence upload and running pipeline

See also

Examples

data('aotus')
seqrec <- aotus@sqs@sqs[[1]]
# this is a SeqRec object
# it contains sequence records
show(seqrec)

#> SeqRec [ID: FJ623078.1]

# you can access its different data slots with @
seqrec@id       # sequence ID, accession + feature location

#> [1] "FJ623078.1"

seqrec@nm       # feature name, '' if none

#> [1] ""

seqrec@accssn   # accession

#> [1] "FJ623078"

seqrec@vrsn     # accession version

#> [1] "FJ623078.1"

seqrec@url      # NCBI GenBank URL

#> [1] "https://www.ncbi.nlm.nih.gov/nuccore/FJ623078.1"

seqrec@txid     # Taxonomic ID

#> [1] "37293"

seqrec@orgnsm   # free-text organism name

#> [1] "Aotus nancymaae"

seqrec@sq       # sequence, in raw format

#>    [1] 61 74 67 67 61 74 67 67 67 67 67 61 61 74 63 63 63 74 74 74 63 63 67 67
#>   [25] 63 61 63 74 74 67 63 74 63 63 74 67 67 74 67 63 74 67 63 61 61 63 74 67
#>   [49] 67 63 67 63 74 61 63 74 63 63 63 61 67 63 61 67 74 63 61 63 74 63 61 63
#>   [73] 67 67 61 61 61 67 61 63 61 67 74 67 67 74 67 63 74 67 67 67 63 67 61 61
#>   [97] 61 61 61 67 67 67 67 61 67 61 63 67 67 74 67 67 61 67 63 74 67 63 63 63
#>  [121] 74 67 74 67 61 61 61 63 74 74 63 63 63 74 67 61 61 67 61 61 67 61 61 63
#>  [145] 67 74 61 63 61 61 74 74 63 63 61 63 74 67 67 61 61 61 61 63 63 74 63 63
#>  [169] 67 61 63 63 61 67 61 74 61 61 61 67 61 74 74 63 74 67 67 67 61 61 61 74
#>  [193] 63 61 67 67 67 63 74 63 63 74 74 63 74 74 67 61 63 74 61 67 61 67 67 74
#>  [217] 63 61 61 74 63 63 61 61 67 63 74 67 67 63 63 67 61 74 63 67 63 61 74 74
#>  [241] 67 61 63 74 63 61 61 61 67 61 61 61 61 67 63 74 63 74 74 67 67 67 61 63
#>  [265] 63 67 61 67 67 61 74 63 63 74 74 74 63 63 63 63 74 67 61 74 63 61 74 63
#>  [289] 61 61 67 61 61 74 67 74 74 63 61 67 67 74 61 67 61 61 67 61 63 74 63 61
#>  [313] 67 61 67 61 63 74 74 61 63 61 74 63 74 67 74 67 61 61 67 74 67 67 61 67
#>  [337] 61 67 61 61 61 67 61 61 67 67 61 67 67 61 67 67 74 67 67 61 61 74 74 67
#>  [361] 63 61 67 67 74 67 74 74 63 67 67 61 74 74 67 61 63 74 67 63 63 61 67 63
#>  [385] 63 63 74 67 61 63 61 63 63 63 61 63 63 74 67 63 74 74 63 61 67 67 67 67
#>  [409] 63 61 67 61 67 63 63 74 67 61 63 63 63 74 67 61 63 63 74 74 67 67 61 67
#>  [433] 61 67 63 63 63 63 63 63 74 67 67 74 61 67 74 61 67 63 63 63 63 74 63 61
#>  [457] 67 74 67 67 61 61 74 67 63 61 63 67 61 67 74 63 63 61 63 67 67 67 67 74
#>  [481] 61 61 61 61 67 67 61 74 61 63 61 67 67 67 67 61 74 67 61 61 67 61 63 63
#>  [505] 63 74 74 74 63 74 67 74 61 74 63 74 63 61 67 63 74 67 67 61 67 61 74 63
#>  [529] 63 61 67 67 61 74 61 67 74 67 67 63 61 63 63 74 67 67 61 61 61 74 67 63
#>  [553] 61 63 63 67 74 63 74 63 63 63 61 67 63 61 63 63 63 67 67 61 67 63 74 67
#>  [577] 74 74 67 74 74 63 61 61 61 61 74 61 61 61 63 67 74 63 67 74 67 67 74 61
#>  [601] 63 74 61 67 63 74 74 74 63 63 61 67 63 61 67 67 63 63 74 63 63 61 67 63
#>  [625] 61 63 61 67 74 67 74 61 74 61 61 67 61 61 61 67 61 67 67 67 67 67 61 61
#>  [649] 63 61 67 67 74 67 67 61 67 74 74 63 74 63 63 74 74 63 63 63 61 63 74 63
#>  [673] 67 63 63 74 74 74 67 63 61 67 63 74 67 61 61 61 61 67 63 74 67 61 63 67
#>  [697] 67 67 63 61 67 74 67 67 63 67 61 67 63 74 67 74 67 63 74 67 67 63 61 67
#>  [721] 67 63 67 67 61 67 61 61 67 67 63 74 74 63 63 74 63 63 74 63 63 61 61 67
#>  [745] 74 63 63 74 67 67 61 74 63 74 63 63 74 74 63 61 61 63 63 74 67 61 63 67
#>  [769] 61 67 63 63 61 67 67 61 67 67 74 67 74 67 74 67 74 61 61 61 61 67 74 67
#>  [793] 67 74 74 61 63 63 63 61 67 67 61 63 63 63 63 61 61 67 63 74 63 61 67 67
#>  [817] 61 74 67 67 67 63 61 61 67 61 61 67 63 74 74 63 63 61 63 74 74 63 61 63
#>  [841] 63 74 63 61 63 63 63 74 67 63 63 63 63 61 67 67 63 63 74 74 67 63 63 74
#>  [865] 63 61 67 74 61 74 67 63 74 67 67 63 74 63 63 67 67 61 61 61 63 74 74 63
#>  [889] 61 63 63 63 74 67 67 63 74 63 74 74 61 61 61 67 67 67 61 61 61 61 74 67
#>  [913] 67 67 61 61 61 67 74 74 67 63 61 74 63 61 67 61 61 61 67 74 67 61 61 63
#>  [937] 63 74 74 67 74 67 67 74 67 61 74 67 61 67 61 67 63 61 61 63 74 63 61 67
#>  [961] 63 74 63 63 61 67 61 61 63 61 61 74 74 74 67 61 63 63 74 67 74 67 61 67
#>  [985] 67 74 67 74 67 67 67 67 61 63 63 63 61 63 63 74 63 63 63 63 74 61 61 67
#> [1009] 63 74 67 61 74 67 63 74 67 61 67 63 74 74 67 61 61 61 63 74 67 67 61 67
#> [1033] 61 61 63 63 61 67 67 61 67 67 63 61 61 61 67 67 74 63 74 63 63 61 61 67
#> [1057] 63 61 67 67 61 67 61 61 67 67 63 67 67 74 67 74 67 67 67 74 67 63 74 67
#> [1081] 61 61 63 63 63 74 67 61 67 67 63 67 67 67 67 67 74 67 74 67 67 63 61 67
#> [1105] 74 67 74 63 74 67 63 74 67 61 67 74 67 61 63 74 63 61 67 67 61 63 61 67
#> [1129] 67 74 63 63 74 67 63 74 67 67 61 61 74 63 63 61 61 67 67 74 63 67 61 67
#> [1153] 67 74 74 63 74 67 63 63 63 61 63 61 74 67 67 74 63 74 63 63 63 63 63 67
#> [1177] 67 74 67 63 61 67 63 63 61 61 74 67 67 63 63 63 74 67 61 74 74 67 74 67
#> [1201] 63 63 67 67 67 67 67 67 74 67 74 63 67 63 67 67 67 63 63 74 63 63 74 67
#> [1225] 67 74 74 74 74 74 61 63 74 67 67 67 63 74 61 67 67 63 61 74 63 74 74 63
#> [1249] 74 74 63 74 67 74 67 74 63 61 67 61 74 67 63 63 67 67 63 61 74 63 67 61
#> [1273] 61 67 67 63 67 63 63 61 61 67 63 61 67 61 67 63 67 67 61 74 67 74 63 74
#> [1297] 63 61 67 61 74 63 61 61 67 61 67 61 63 74 74 63 74 63 61 67 74 67 61 67
#> [1321] 61 61 67 61 61 67 61 63 63 74 67 63 63 61 67 74 67 63 63 63 63 63 61 63
#> [1345] 63 67 67 74 74 74 63 61 67 61 61 67 61 63 61 74 67 74 61 67 63 63 63 63
#> [1369] 61 74 74 74 67 61

seqrec@dfln     # sequence definition

#> [1] "Aotus nancymaae CD4 antigen (CD4) mRNA, complete cds"

seqrec@ml_typ   # molecule type

#> [1] "mRNA"

seqrec@rec_typ  # whole record or feature

#> [1] "whole"

seqrec@nncltds  # sequence length

#> [1] 1374

seqrec@nambgs   # number of non-ATCGs

#> [1] 0

seqrec@pambgs   # proportion of non-ATCGs

#> [1] 0

seqrec@gcr      # GC-ratio

#> [1] 0.4468705

seqrec@age      # days since being added to GenBank

#> [1] 3412

# get the sequence like so....
(rawToChar(seqrec@sq))

#> [1] "atggatgggggaatccctttccggcacttgctcctggtgctgcaactggcgctactcccagcagtcactcacggaaagacagtggtgctgggcgaaaaaggggagacggtggagctgccctgtgaaacttccctgaagaagaacgtacaattccactggaaaacctccgaccagataaagattctgggaaatcagggctccttcttgactagaggtcaatccaagctggccgatcgcattgactcaaagaaaagctcttgggaccgaggatcctttcccctgatcatcaagaatgttcaggtagaagactcagagacttacatctgtgaagtggagagaaagaaggaggaggtggaattgcaggtgttcggattgactgccagccctgacacccacctgcttcaggggcagagcctgaccctgaccttggagagcccccctggtagtagcccctcagtggaatgcacgagtccacggggtaaaaggatacaggggatgaagaccctttctgtatctcagctggagatccaggatagtggcacctggaaatgcaccgtctcccagcacccggagctgttgttcaaaataaacgtcgtggtactagctttccagcaggcctccagcacagtgtataagaaagagggggaacaggtggagttctccttcccactcgcctttgcagctgaaaagctgacgggcagtggcgagctgtgctggcaggcggagaaggcttcctcctccaagtcctggatctccttcaacctgacgagccaggaggtgtgtgtaaaagtggttacccaggaccccaagctcaggatgggcaagaagcttccacttcacctcaccctgccccaggccttgcctcagtatgctggctccggaaacttcaccctggctcttaaagggaaaatgggaaagttgcatcagaaagtgaaccttgtggtgatgagagcaactcagctccagaacaatttgacctgtgaggtgtggggacccacctcccctaagctgatgctgagcttgaaactggagaaccaggaggcaaaggtctccaagcaggagaaggcggtgtgggtgctgaaccctgaggcgggggtgtggcagtgtctgctgagtgactcaggacaggtcctgctggaatccaaggtcgaggttctgcccacatggtctcccccggtgcagccaatggccctgattgtgccggggggtgtcgcgggcctcctggtttttactgggctaggcatcttcttctgtgtcagatgccggcatcgaaggcgccaagcagagcggatgtctcagatcaagagacttctcagtgagaagaagacctgccagtgcccccaccggtttcagaagacatgtagccccatttga"