Phylota object

Phylota table contains all sequence, cluster and taxonomic information from a phylotaR pipeline run.

Usage

# S4 method for Phylota
as.character(x)

# S4 method for Phylota
show(object)

# S4 method for Phylota
print(x)

# S4 method for Phylota
str(object, max.level = 2L, ...)

# S4 method for Phylota
summary(object)

# S4 method for Phylota,character
[[(x, i)

Arguments

x

Phylota object

object

Phylota object

max.level

Maximum level of nesting for str()

...

Further arguments for str()

i

Either sid or cid

Slots

cids

IDs of all clusters

sids

IDs of all sequences

txids

IDs of all taxa

sqs

All sequence records as SeqArc

clstrs

All cluster records as ClstrArc

txdct

Taxonomic dictionary as TaxDict

prnt_id

Parent taxonomic ID

prnt_nm

Parent taxonomic name

See also

Examples

data('aotus')
# this is a Phylota object
# it contains cluster, sequence and taxonomic information from a phylotaR run
show(aotus)
#> Phylota Table (Aotus)
#> - [184] clusters
#> - [1945] sequences
#> - [13] source taxa
# you can access its different data slots with @
aotus@cids   # cluster IDs
#>   [1] "0"   "1"   "2"   "3"   "4"   "5"   "6"   "7"   "8"   "9"   "10"  "11" 
#>  [13] "12"  "13"  "14"  "15"  "16"  "17"  "18"  "19"  "20"  "21"  "22"  "23" 
#>  [25] "24"  "25"  "26"  "27"  "28"  "29"  "30"  "31"  "32"  "33"  "34"  "35" 
#>  [37] "36"  "37"  "38"  "39"  "40"  "41"  "42"  "43"  "44"  "45"  "46"  "47" 
#>  [49] "48"  "49"  "50"  "51"  "52"  "53"  "54"  "55"  "56"  "57"  "58"  "59" 
#>  [61] "60"  "61"  "62"  "63"  "64"  "65"  "66"  "67"  "68"  "69"  "70"  "71" 
#>  [73] "72"  "73"  "74"  "75"  "76"  "77"  "78"  "79"  "80"  "81"  "82"  "83" 
#>  [85] "84"  "85"  "86"  "87"  "88"  "89"  "90"  "91"  "92"  "93"  "94"  "95" 
#>  [97] "96"  "97"  "98"  "99"  "100" "101" "102" "103" "104" "105" "106" "107"
#> [109] "108" "109" "110" "111" "112" "113" "114" "115" "116" "117" "118" "119"
#> [121] "120" "121" "122" "123" "124" "125" "126" "127" "128" "129" "130" "131"
#> [133] "132" "133" "134" "135" "136" "137" "138" "139" "140" "141" "142" "143"
#> [145] "144" "145" "146" "147" "148" "149" "150" "151" "152" "153" "154" "155"
#> [157] "156" "157" "158" "159" "160" "161" "162" "163" "164" "165" "166" "167"
#> [169] "168" "169" "170" "171" "172" "173" "174" "175" "176" "177" "178" "179"
#> [181] "180" "181" "182" "183"
aotus@sids   # sequence IDs
#>    [1] "KR528418.1" "JF444279.1" "AY227074.1" "AY227073.1" "AY227072.1"
#>    [6] "AY227071.1" "AY227070.1" "AY227069.1" "AY227068.1" "AY227067.1"
#>   [11] "AY227066.1" "AY227065.1" "AY227064.1" "AY227063.1" "AY227062.1"
#>   [16] "AY227061.1" "AY227060.1" "AY227059.1" "AY227058.1" "AY227057.1"
#>   [21] "AY227056.1" "AY227055.1" "DQ321669.1" "MW327030.1" "MT489117.1"
#>   [26] "MT489116.1" "MT489115.1" "MT489114.1" "MT489113.1" "MT489112.1"
#>   [31] "MT489111.1" "MT489110.1" "MT489075.1" "MT489074.1" "MT489073.1"
#>   [36] "MT489072.1" "MT489071.1" "MT489070.1" "MT489069.1" "MT489068.1"
#>   [41] "MT489032.1" "MT489031.1" "MT489030.1" "MT489029.1" "MT489028.1"
#>   [46] "MT489027.1" "MT489026.1" "MT489025.1" "MT488990.1" "MT488989.1"
#>   [51] "MT488988.1" "MT488987.1" "MT488986.1" "MT488985.1" "MT488984.1"
#>   [56] "MT488983.1" "MT488947.1" "MT488946.1" "MT488945.1" "MT488944.1"
#>   [61] "MT488943.1" "MT488942.1" "MT488941.1" "MT488940.1" "MT488904.1"
#>   [66] "MT488903.1" "MT488902.1" "MT488901.1" "MT488900.1" "MT488899.1"
#>   [71] "MT488898.1" "MT488897.1" "MT488861.1" "MT488860.1" "MT488859.1"
#>   [76] "MT488858.1" "MT488857.1" "MT488856.1" "MT488855.1" "MT488854.1"
#>   [81] "MT488815.1" "MT488814.1" "MT488813.1" "MT488812.1" "MT488811.1"
#>   [86] "MT488810.1" "MT488809.1" "MT488808.1" "MT488778.1" "MT488777.1"
#>   [91] "MT488776.1" "MT488775.1" "MT488774.1" "MT488773.1" "MT488772.1"
#>   [96] "MT488771.1" "MT488734.1" "MT488733.1" "MT488732.1" "MT488731.1"
#>  [101] "MT488730.1" "MT488729.1" "MT488728.1" "MT488727.1" "MT488672.1"
#>  [106] "MT488671.1" "MT488670.1" "MT488669.1" "MT488659.1" "MT488658.1"
#>  [111] "MT488657.1" "MT488569.1" "MT488568.1" "MT488567.1" "MT488566.1"
#>  [116] "MF632281.1" "MF632280.1" "KX774456.1" "KR902344.1" "DQ321666.1"
#>  [121] "DQ321665.1" "DQ162635.1" "DQ162634.1" "DQ162633.1" "DQ162630.1"
#>  [126] "DQ162629.1" "DQ162628.1" "DQ162627.1" "DQ162626.1" "DQ162624.1"
#>  [131] "AY563263.2" "AY563258.2" "AY563227.2" "AY563219.2" "AY659834.1"
#>  [136] "AY659833.1" "AY659832.1" "AY659831.1" "AY659830.1" "AY659829.1"
#>  [141] "AF352260.1" "AF352259.1" "AF014508.1" "AF014507.1" "AF014506.1"
#>  [146] "AF014504.1" "HQ005503.1" "HQ005502.1" "HQ018928.1" "AY114539.1"
#>  [151] "AY114538.1" "AY114537.1" "AY114536.1" "AY114535.1" "AY114534.1"
#>  [156] "AY114533.1" "AY114532.1" "AY114531.1" "AY114530.1" "AY114529.1"
#>  [161] "KC762196.1" "KC762163.1" "KC762106.1" "KC762075.1" "KC762017.1"
#>  [166] "KC761983.1" "KC761954.1" "KC761921.1" "KC761892.1" "KC761836.1"
#>  [171] "KC761816.1" "KC761785.1" "KC761698.1" "KC761635.1" "KC761604.1"
#>  [176] "KC761572.1" "KC761540.1" "KC761507.1" "KC761473.1" "KC761408.1"
#>  [181] "KC761373.1" "KC761279.1" "KC761245.1" "KC761212.1" "KC761180.1"
#>  [186] "KC761083.1" "KC761053.1" "KC761019.1" "KC760991.1" "MW736163.1"
#>  [191] "MT489140.1" "MT489139.1" "MT489137.1" "MT489136.1" "MT489135.1"
#>  [196] "MT489134.1" "MT489095.1" "MT489094.1" "MT489092.1" "MT489091.1"
#>  [201] "MT489053.1" "MT489052.1" "MT489050.1" "MT489049.1" "MT489048.1"
#>  [206] "MT489010.1" "MT489009.1" "MT489007.1" "MT489006.1" "MT488968.1"
#>  [211] "MT488967.1" "MT488965.1" "MT488964.1" "MT488963.1" "MT488925.1"
#>  [216] "MT488924.1" "MT488922.1" "MT488921.1" "MT488920.1" "MT488882.1"
#>  [221] "MT488881.1" "MT488879.1" "MT488878.1" "MT488839.1" "MT488838.1"
#>  [226] "MT488800.1" "MT488799.1" "MT488797.1" "MT488796.1" "MT488795.1"
#>  [231] "MT488756.1" "MT488755.1" "MT488753.1" "MT488752.1" "MT488751.1"
#>  [236] "MT488681.1" "MT488680.1" "MT488679.1" "MT488678.1" "MT488677.1"
#>  [241] "MT488676.1" "MT488592.1" "MT488591.1" "MT488589.1" "MT488588.1"
#>  [246] "MT488587.1" "MT488586.1" "MT492189.1" "MT492188.1" "MT483740.1"
#>  [251] "MT483739.1" "MT483737.1" "MT483736.1" "MT483735.1" "MT483734.1"
#>  [256] "MF615381.1" "MF615380.1" "MF615379.1" "MF615378.1" "MF615377.1"
#>  [261] "MF615376.1" "MK205145.1" "MF632285.1" "MF632284.1" "MF632283.1"
#>  [266] "MF632282.1" "MW736164.1" "MT489147.1" "MT489141.1" "MT489126.1"
#>  [271] "MT489106.1" "MT489102.1" "MT489096.1" "MT489083.1" "MT489064.1"
#>  [276] "MT489060.1" "MT489054.1" "MT489040.1" "MT489021.1" "MT489017.1"
#>  [281] "MT489011.1" "MT488998.1" "MT488979.1" "MT488975.1" "MT488969.1"
#>  [286] "MT488955.1" "MT488936.1" "MT488926.1" "MT488912.1" "MT488893.1"
#>  [291] "MT488889.1" "MT488883.1" "MT488870.1" "MT488850.1" "MT488846.1"
#>  [296] "MT488840.1" "MT488831.1" "MT488807.1" "MT488801.1" "MT488787.1"
#>  [301] "MT488767.1" "MT488763.1" "MT488757.1" "MT488743.1" "MT488723.1"
#>  [306] "MT488675.1" "MT488655.1" "MT488599.1" "MT488593.1" "MT488578.1"
#>  [311] "MT488558.1" "MT492196.1" "MT492190.1" "MT492179.1" "MT483747.1"
#>  [316] "MT483741.1" "MT483726.1" "MT483706.1" "DQ321664.1" "DQ162736.1"
#>  [321] "DQ162732.1" "DQ162729.1" "DQ162727.1" "DQ162719.1" "DQ162718.1"
#>  [326] "DQ162711.1" "DQ162710.1" "AY563262.2" "AY563261.2" "AY563260.2"
#>  [331] "AY563259.2" "AY563257.2" "AY563256.2" "AY563254.2" "AY563253.2"
#>  [336] "AY563252.2" "AY563251.2" "AY563250.2" "AY563249.2" "AY563248.2"
#>  [341] "AY563247.2" "AY563246.2" "AY563245.2" "AY563244.2" "AY563243.2"
#>  [346] "AY563242.2" "AY563241.2" "AY563229.2" "AY563226.2" "AY659828.1"
#>  [351] "AY659827.1" "AY659826.1" "AF352258.1" "AF352257.1" "AF352256.1"
#>  [356] "AF132757.1" "AF129797.1" "AF097328.1" "AF097326.1" "AF097324.1"
#>  [361] "JQ932319.1" "JQ932318.1" "JQ932310.1" "JQ932304.1" "JQ932303.1"
#>  [366] "JQ932293.1" "JQ932280.1" "JQ932259.1" "JQ932255.1" "JQ932245.1"
#>  [371] "JQ932232.1" "JQ932362.1" "JQ932359.1" "JQ932356.1" "JQ932605.1"
#>  [376] "JQ932597.1" "JQ932590.1" "JQ932717.1" "JQ932662.1" "JQ932636.1"
#>  [381] "JQ932630.1" "JQ932822.1" "JQ932821.1" "JQ932818.1" "JQ932817.1"
#>  [386] "JQ932816.1" "JQ932815.1" "JQ932814.1" "JQ932813.1" "JQ932812.1"
#>  [391] "JQ932809.1" "JQ932807.1" "JQ932806.1" "JQ932805.1" "JQ932804.1"
#>  [396] "JQ932803.1" "JQ932802.1" "JQ932801.1" "JQ932800.1" "JQ932799.1"
#>  [401] "JQ932794.1" "JQ932775.1" "JQ932756.1" "JQ932751.1" "JQ932724.1"
#>  [406] "JQ932932.1" "JQ932931.1" "JQ932930.1" "JQ932929.1" "JQ932928.1"
#>  [411] "JQ932927.1" "JQ932926.1" "JQ932925.1" "JQ932924.1" "JQ932923.1"
#>  [416] "JQ932922.1" "JQ932921.1" "JQ932920.1" "JQ932918.1" "JQ932916.1"
#>  [421] "JQ932914.1" "JQ932913.1" "JQ932909.1" "JQ932908.1" "JQ932905.1"
#>  [426] "JQ932903.1" "JQ932897.1" "JQ932896.1" "JQ932895.1" "JQ932892.1"
#>  [431] "JQ932891.1" "JQ932890.1" "JQ932889.1" "JQ932888.1" "JQ932887.1"
#>  [436] "JQ932886.1" "JQ932885.1" "JQ932884.1" "JQ932882.1" "JQ932881.1"
#>  [441] "JQ932880.1" "JQ932879.1" "JQ932877.1" "JQ932874.1" "JQ932873.1"
#>  [446] "JQ932872.1" "JQ932870.1" "JQ932869.1" "JQ932868.1" "JQ932867.1"
#>  [451] "JQ932866.1" "JQ932865.1" "JQ932864.1" "JQ932863.1" "JQ932862.1"
#>  [456] "JQ932861.1" "JQ932859.1" "JQ932857.1" "JQ932856.1" "JQ932855.1"
#>  [461] "JQ932854.1" "JQ932853.1" "JQ932852.1" "JQ932851.1" "JQ932850.1"
#>  [466] "JQ932849.1" "JQ932847.1" "JQ932846.1" "JQ932845.1" "JQ932844.1"
#>  [471] "JQ932843.1" "JQ932842.1" "JQ932841.1" "JQ932840.1" "JQ932839.1"
#>  [476] "JQ932838.1" "JQ932837.1" "JQ932836.1" "JQ932835.1" "JQ932833.1"
#>  [481] "JQ932831.1" "JQ932830.1" "JQ932829.1" "JQ932827.1" "JQ932823.1"
#>  [486] "JQ933029.1" "JQ933028.1" "JQ933027.1" "JQ933026.1" "JQ933023.1"
#>  [491] "JQ933020.1" "JQ933011.1" "JQ933006.1" "JQ933005.1" "JQ933004.1"
#>  [496] "JQ933001.1" "JQ933000.1" "JQ932999.1" "JQ932998.1" "JQ932997.1"
#>  [501] "JQ932996.1" "JQ932995.1" "JQ932994.1" "JQ932993.1" "JQ932992.1"
#>  [506] "JQ932991.1" "JQ932990.1" "JQ932989.1" "JQ932988.1" "JQ932987.1"
#>  [511] "JQ932986.1" "JQ932985.1" "JQ932984.1" "JQ932983.1" "JQ932982.1"
#>  [516] "JQ932981.1" "JQ932980.1" "JQ932979.1" "JQ932978.1" "JQ932977.1"
#>  [521] "JQ932976.1" "JQ932975.1" "JQ932974.1" "JQ932973.1" "JQ932972.1"
#>  [526] "JQ932971.1" "JQ932970.1" "JQ932969.1" "JQ932967.1" "JQ932966.1"
#>  [531] "JQ932965.1" "JQ932964.1" "JQ932963.1" "JQ932962.1" "JQ932961.1"
#>  [536] "JQ932960.1" "JQ932959.1" "JQ932958.1" "JQ932957.1" "JQ932956.1"
#>  [541] "JQ932955.1" "JQ932954.1" "JQ932953.1" "JQ932934.1" "JQ932933.1"
#>  [546] "JQ963361.1" "JQ963360.1" "JQ963351.1" "JQ963350.1" "JQ963349.1"
#>  [551] "JQ963348.1" "JQ963347.1" "JQ963346.1" "JQ933101.1" "JQ933099.1"
#>  [556] "JQ933082.1" "JQ933076.1" "JQ933075.1" "JQ933053.1" "MT489118.1"
#>  [561] "MT489076.1" "MT489033.1" "MT488991.1" "MT488948.1" "MT488905.1"
#>  [566] "MT488862.1" "MT488816.1" "MT488779.1" "MT488735.1" "MT488656.1"
#>  [571] "MT488570.1" "MT492172.1" "MT483718.1" "MT253565.1" "MT232916.1"
#>  [576] "MT232915.1" "LR699078.1" "LR699077.1" "LR699076.1" "LS481196.1"
#>  [581] "LS481195.1" "LS481194.1" "LS481193.1" "LS481192.1" "AF097329.1"
#>  [586] "AF097327.1" "AF097325.1" "AF097321.1" "JN161047.1" "HQ005506.1"
#>  [591] "HQ018927.1" "KC762193.1" "KC762160.1" "KC762127.1" "KC762104.1"
#>  [596] "KC762072.1" "KC762014.1" "KC761980.1" "KC761951.1" "KC761918.1"
#>  [601] "KC761889.1" "KC761833.1" "KC761813.1" "KC761782.1" "KC761755.1"
#>  [606] "KC761695.1" "KC761632.1" "KC761601.1" "KC761569.1" "KC761537.1"
#>  [611] "KC761504.1" "KC761470.1" "KC761438.1" "KC761405.1" "KC761371.1"
#>  [616] "KC761338.1" "KC761242.1" "KC761210.1" "KC761177.1" "KC761144.1"
#>  [621] "KC761111.1" "KC761080.1" "KC761050.1" "KC761016.1" "KC760988.1"
#>  [626] "KC760954.1" "KC760891.1" "KC760866.1" "KC760833.1" "KC760802.1"
#>  [631] "KC760770.1" "KC760738.1" "KC760712.1" "KC760681.1" "KC760646.1"
#>  [636] "KC760611.1" "KC760576.1" "KC760514.1" "KC760483.1" "KC760450.1"
#>  [641] "KC760417.1" "KC760385.1" "KC760326.1" "KC760263.1" "KC760228.1"
#>  [646] "KC761952.1" "KC761919.1" "KC761890.1" "KC761834.1" "KC761814.1"
#>  [651] "KC761783.1" "KC761756.1" "KC761696.1" "KC761633.1" "KC761602.1"
#>  [656] "KC761570.1" "KC761538.1" "KC761505.1" "KC761471.1" "KC761439.1"
#>  [661] "KC761406.1" "KC761339.1" "KC761277.1" "KC761243.1" "KC761178.1"
#>  [666] "KC761145.1" "KC761081.1" "KC761051.1" "KC761017.1" "KC760989.1"
#>  [671] "KC760955.1" "KC760892.1" "KC760867.1" "KC760834.1" "KC760803.1"
#>  [676] "KC760771.1" "KC760739.1" "KC760713.1" "KC760682.1" "KC760647.1"
#>  [681] "KC760612.1" "KC760577.1" "KC760515.1" "KC760484.1" "KC760451.1"
#>  [686] "KC760418.1" "KC760386.1" "KC760327.1" "KC760264.1" "KC760229.1"
#>  [691] "AY225808.2" "AY225809.2" "AF169487.1" "AF169486.1" "AF169485.1"
#>  [696] "U36770.1"   "HM756840.1" "HM757383.1" "HM757558.1" "HM757724.1"
#>  [701] "HM758216.1" "HM758297.1" "HM758465.1" "HM758650.1" "HM758934.1"
#>  [706] "HM759105.1" "HM759278.1" "HM759372.1" "HM759646.1" "HM759729.1"
#>  [711] "HM759896.1" "HM760080.1" "HM760256.1" "HM760529.1" "HM760608.1"
#>  [716] "HM760775.1" "HM761013.1" "HM761145.1" "HM761195.1" "HM761507.1"
#>  [721] "HM761772.1" "HM761929.1" "HM762096.1" "HM762207.1" "HM762431.1"
#>  [726] "HM762511.1" "HM762768.1" "HM762858.1" "HM763021.1" "HM763265.1"
#>  [731] "HM763433.1" "HM763726.1" "HM763996.1" "HM764104.1" "HM764321.1"
#>  [736] "HM764396.1" "HM764648.1" "AF185631.1" "AF185630.1" "AF185629.1"
#>  [741] "AF185628.1" "AF185627.1" "AF185625.1" "AF185624.1" "AF185623.1"
#>  [746] "AF107758.1" "AF107755.1" "AF107753.1" "AF107750.1" "AF107749.1"
#>  [751] "AF107748.1" "AF107742.1" "AF107741.1" "AF107740.1" "AF107738.1"
#>  [756] "AF107736.1" "AF014513.1" "AF014512.1" "AF014511.1" "AF014509.1"
#>  [761] "JN161054.1" "JN161053.1" "JN161052.1" "JN161051.1" "JN161050.1"
#>  [766] "JN161049.1" "HQ005509.1" "HQ005508.1" "HQ005507.1" "AY114521.1"
#>  [771] "AY114520.1" "AY114519.1" "AY114518.1" "AY114517.1" "AY114516.1"
#>  [776] "AY114515.1" "AY114514.1" "AF027569.1" "AF027567.1" "AF027566.1"
#>  [781] "AF027565.1" "AF027564.1" "AF027563.1" "AF027562.1" "AF027561.1"
#>  [786] "AF027560.1" "AF027559.1" "AF027554.1" "AF027553.1" "AF027552.1"
#>  [791] "AF027551.1" "AF027549.1" "AF027548.1" "AF027547.1" "AF027546.1"
#>  [796] "AF027545.1" "AF027544.1" "AF027543.1" "AF027542.1" "AF286382.1"
#>  [801] "AF286380.1" "KC762194.1" "KC762161.1" "KC762128.1" "KC762073.1"
#>  [806] "KC762015.1" "KC761981.1" "AY245679.1" "AY245678.1" "AY245677.1"
#>  [811] "AY245676.1" "AY245675.1" "AY283262.1" "AY225807.1" "AY225804.1"
#>  [816] "AY225803.1" "AY225799.1" "AF352255.1" "AF352254.1" "AF333732.1"
#>  [821] "AF333731.1" "AF333730.1" "AF333729.1" "AF333728.1" "AF333727.1"
#>  [826] "AF333726.1" "AF333725.1" "AF333724.1" "AF333723.1" "AF333722.1"
#>  [831] "AF333721.1" "AF333720.1" "AF333719.1" "AF333718.1" "AF333717.1"
#>  [836] "AF333716.1" "AF333715.1" "AF333714.1" "AF333713.1" "AF333712.1"
#>  [841] "AF333711.1" "AF333710.1" "AF333709.1" "AF378750.1" "AF378749.1"
#>  [846] "AF378748.1" "AF378747.1" "AF378746.1" "AF378745.1" "AF378744.1"
#>  [851] "AF378743.1" "AF378742.1" "AF378741.1" "AF336946.1" "AF336945.1"
#>  [856] "AF336944.1" "AF336943.1" "AF336942.1" "AF336941.1" "AF336940.1"
#>  [861] "AF336939.1" "AF336938.1" "AF336937.1" "AF336936.1" "AF336935.1"
#>  [866] "AF336934.1" "AF336933.1" "AF336932.1" "AF336931.1" "AF336930.1"
#>  [871] "AF440688.1" "AF132770.1" "AF132769.1" "AF132768.1" "AF132767.1"
#>  [876] "AF132766.1" "AF132765.1" "AF132764.1" "AF132763.1" "AF132762.1"
#>  [881] "AF132761.1" "AF132760.1" "AF132759.1" "AF132758.1" "AF132756.1"
#>  [886] "AF132755.1" "AF129808.1" "AF129807.1" "AF129806.1" "AF129805.1"
#>  [891] "AF129804.1" "AF129803.1" "AF129802.1" "AF129801.1" "AF129800.1"
#>  [896] "AF129799.1" "AF129798.1" "AF129796.1" "AF129795.1" "AF129794.1"
#>  [901] "AF129793.1" "AF129792.1" "AF185633.1" "AF185632.1" "AY563202.2"
#>  [906] "AY563201.2" "AY563200.2" "AY563199.2" "AY563198.2" "AY563197.2"
#>  [911] "AY563196.2" "AY563195.2" "AY563194.2" "AY563193.2" "AY563192.2"
#>  [916] "AY563191.2" "AY563190.2" "AY563189.2" "AY563188.2" "AY563186.2"
#>  [921] "AY563185.2" "AY563184.2" "AY563183.2" "AY563182.2" "AY563181.2"
#>  [926] "AY563180.2" "AY972694.1" "AY900570.1" "AY900566.1" "AY900565.1"
#>  [931] "AY900564.1" "AY900561.1" "AY900560.1" "AY900559.1" "AY900557.1"
#>  [936] "AY900556.1" "AY900554.1" "AY900553.1" "AY900552.1" "AY900550.1"
#>  [941] "AY900549.1" "AY900548.1" "AY900547.1" "AY900546.1" "AY900544.1"
#>  [946] "AY900539.1" "AY900538.1" "AY900537.1" "AY900536.1" "AY900535.1"
#>  [951] "AY900534.1" "AY900533.1" "AY900532.1" "AY900531.1" "AY900530.1"
#>  [956] "AY900529.1" "AY900528.1" "AY900527.1" "AY900524.1" "AY900520.1"
#>  [961] "AY659852.1" "AY659851.1" "AY659850.1" "AY659849.1" "AY659848.1"
#>  [966] "AY659847.1" "AY659846.1" "AY659845.1" "AY659844.1" "AY659843.1"
#>  [971] "AY659842.1" "AY659841.1" "AY659840.1" "AY659839.1" "AY659838.1"
#>  [976] "AY659837.1" "AY659836.1" "AY659835.1" "AY493661.1" "AY245688.1"
#>  [981] "AY245687.1" "AY245686.1" "AY245685.1" "AY245684.1" "AY245683.1"
#>  [986] "AY245682.1" "AY245681.1" "AY245680.1" "MT483724.1" "MT483722.1"
#>  [991] "MT483721.1" "MT483720.1" "MT483719.1" "MF615384.1" "MF615383.1"
#>  [996] "MF615382.1" "MF615375.1" "MF615374.1" "MF615373.1" "MF615372.1"
#> [1001] "MF615371.1" "MF632279.1" "KR902343.1" "DQ321660.1" "DQ321659.1"
#> [1006] "DQ162705.1" "DQ162704.1" "DQ162699.1" "DQ162688.1" "DQ162683.1"
#> [1011] "DQ162682.1" "DQ162679.1" "DQ162676.1" "DQ162668.1" "DQ162661.1"
#> [1016] "DQ162660.1" "DQ162648.1" "DQ162647.1" "DQ162646.1" "DQ162645.1"
#> [1021] "AY563255.2" "AY563239.2" "AY563237.2" "AY563236.2" "AY563234.2"
#> [1026] "AY563233.2" "AY563232.2" "AY563231.2" "AY563230.2" "AY563228.2"
#> [1031] "AY563225.2" "AY563224.2" "AY563223.2" "AY563222.2" "AY563221.2"
#> [1036] "AY563220.2" "AY563218.2" "AY563217.2" "AY563216.2" "AY563215.2"
#> [1041] "AY563214.2" "AY563213.2" "AY563212.2" "AY563211.2" "AY563210.2"
#> [1046] "AY563209.2" "AY563208.2" "AY563207.2" "AY563206.2" "AY563204.2"
#> [1051] "AY563203.2" "MW321654.1" "MW321653.1" "MT489138.1" "MT489125.1"
#> [1056] "MT489124.1" "MT489122.1" "MT489121.1" "MT489120.1" "MT489119.1"
#> [1061] "MT489093.1" "MT489082.1" "MT489081.1" "MT489079.1" "MT489078.1"
#> [1066] "MT489077.1" "MT489051.1" "MT489039.1" "MT489038.1" "MT489036.1"
#> [1071] "MT489035.1" "MT489034.1" "MT489008.1" "MT488997.1" "MT488996.1"
#> [1076] "MT488994.1" "MT488993.1" "MT488992.1" "MT488966.1" "MT488954.1"
#> [1081] "MT488953.1" "MT488951.1" "MT488950.1" "MT488949.1" "MT488923.1"
#> [1086] "MT488911.1" "MT488910.1" "MT488908.1" "MT488907.1" "MT488906.1"
#> [1091] "MT488880.1" "MT488869.1" "MT488868.1" "MT488866.1" "MT488865.1"
#> [1096] "MT488864.1" "MT488863.1" "MT488837.1" "MT488828.1" "MT488827.1"
#> [1101] "MT488825.1" "MT488824.1" "MT488823.1" "MT488798.1" "MT488786.1"
#> [1106] "MT488785.1" "MT488783.1" "MT488782.1" "MT488781.1" "MT488780.1"
#> [1111] "MT488754.1" "MT488742.1" "MT488741.1" "MT488739.1" "MT488738.1"
#> [1116] "MT488737.1" "MT488736.1" "MT488654.1" "MT488653.1" "MT488652.1"
#> [1121] "MT488650.1" "MT488649.1" "MT488648.1" "MT488590.1" "MT488577.1"
#> [1126] "MT488576.1" "MT488574.1" "MT488573.1" "MT488572.1" "MT488571.1"
#> [1131] "MT492187.1" "MT492178.1" "MT492177.1" "MT492175.1" "MT492174.1"
#> [1136] "MT492173.1" "MT483738.1" "MT483725.1" "HM762767.1" "HM762766.1"
#> [1141] "HM762765.1" "HM762863.1" "HM763019.1" "HM763263.1" "HM763262.1"
#> [1146] "HM763261.1" "HM763431.1" "HM763430.1" "HM763429.1" "HM763724.1"
#> [1151] "HM763723.1" "HM763722.1" "HM763994.1" "HM764188.1" "HM764187.1"
#> [1156] "HM764102.1" "HM764319.1" "HM764318.1" "HM764317.1" "HM764402.1"
#> [1161] "HM764646.1" "HM764645.1" "HM764814.1" "HM764813.1" "HM765068.1"
#> [1166] "HM764910.1" "HM764909.1" "HM765177.1" "HM765176.1" "HM765175.1"
#> [1171] "AF420045.1" "AY449080.1" "AY449079.1" "AY449071.1" "AY449070.1"
#> [1176] "AY449069.1" "AY449065.1" "AY449059.1" "AY449058.1" "AY449057.1"
#> [1181] "HM756917.1" "HM756916.1" "HM756838.1" "HM757001.1" "HM757000.1"
#> [1186] "HM756999.1" "HM757381.1" "HM757380.1" "HM757379.1" "HM757556.1"
#> [1191] "HM757555.1" "HM757554.1" "HM757722.1" "HM757721.1" "HM757720.1"
#> [1196] "HM757999.1" "HM758214.1" "HM758213.1" "HM758212.1" "HM758450.1"
#> [1201] "HM758304.1" "HM758303.1" "HM758476.1" "HM758475.1" "HM758463.1"
#> [1206] "HM758660.1" "HM758659.1" "HM758658.1" "HM758657.1" "HM758656.1"
#> [1211] "HM758932.1" "HM758931.1" "HM758930.1" "HM759103.1" "HM759102.1"
#> [1216] "HM759101.1" "HM759276.1" "HM759275.1" "HM759274.1" "HM759273.1"
#> [1221] "HM759535.1" "HM759379.1" "HM759378.1" "HM759644.1" "HM759643.1"
#> [1226] "HM759641.1" "HM759836.1" "HM759777.1" "HM759775.1" "HM760059.1"
#> [1231] "HM759903.1" "HM759902.1" "HM760239.1" "HM760087.1" "HM760086.1"
#> [1236] "HM760421.1" "HM760263.1" "HM760262.1" "HM760527.1" "HM760526.1"
#> [1241] "HM760525.1" "HM760615.1" "HM760614.1" "HM760759.1" "HM761011.1"
#> [1246] "HM761143.1" "HM761207.1" "HM761206.1" "HM761193.1" "HM761664.1"
#> [1251] "HM761514.1" "HM761513.1" "HM761770.1" "HM761769.1" "HM761768.1"
#> [1256] "HM761927.1" "HM761926.1" "HM761925.1" "HM762094.1" "HM762093.1"
#> [1261] "HM762092.1" "HM762324.1" "HM762181.1" "HM762180.1" "HM762429.1"
#> [1266] "HM762428.1" "HM762427.1" "HM762674.1" "HM762517.1" "DQ098868.1"
#> [1271] "DQ098867.1" "DQ098856.1" "DQ098855.1" "AY455159.1" "AY455158.1"
#> [1276] "AY455153.1" "U36846.1"   "JN161098.1" "JN161097.1" "JN161096.1"
#> [1281] "JN161095.1" "JN161094.1" "JN161093.1" "JN161092.1" "JN161091.1"
#> [1286] "JN161090.1" "JN161089.1" "JN161088.1" "JN161087.1" "JN161086.1"
#> [1291] "JN161085.1" "JN161084.1" "JN161083.1" "JN161082.1" "JN161081.1"
#> [1296] "JN161080.1" "JN161079.1" "JN161078.1" "JN161077.1" "JN161076.1"
#> [1301] "JN161075.1" "JN161074.1" "JN161073.1" "JN161072.1" "JN161071.1"
#> [1306] "JN161070.1" "JN161069.1" "EF658653.1" "EF658652.1" "DQ989366.1"
#> [1311] "KC762195.1" "KC762162.1" "KC762129.1" "KC762105.1" "KC762074.1"
#> [1316] "KC762016.1" "KC761982.1" "KC761953.1" "KC761920.1" "KC761891.1"
#> [1321] "KC761835.1" "KC761815.1" "KC761784.1" "KC761757.1" "KC761697.1"
#> [1326] "KC761634.1" "KC761603.1" "KC761571.1" "KC761539.1" "KC761506.1"
#> [1331] "KC761472.1" "KC761440.1" "KC761407.1" "KC761372.1" "KC761340.1"
#> [1336] "KC761278.1" "KC761244.1" "KC761211.1" "KC761179.1" "KC761146.1"
#> [1341] "KC761112.1" "KC761082.1" "KC761052.1" "KC761018.1" "KC760990.1"
#> [1346] "KC760956.1" "KC760893.1" "KC760868.1" "KC760835.1" "KC760804.1"
#> [1351] "KC760772.1" "KC760740.1" "KC760714.1" "KC760683.1" "KC760648.1"
#> [1356] "KC760613.1" "KC760578.1" "KC760516.1" "KC760485.1" "KC760452.1"
#> [1361] "KC760419.1" "KC760387.1" "KC760328.1" "KC760265.1" "KC760230.1"
#> [1366] "AB762011.1" "AB762010.1" "AB762009.1" "AB762008.1" "AB762007.1"
#> [1371] "AB762006.1" "AB762005.1" "AB762004.1" "AB762003.1" "AB762002.1"
#> [1376] "AB762001.1" "AB762000.1" "AB761999.1" "AB761998.1" "AB761997.1"
#> [1381] "AB761996.1" "AB761995.1" "AB761994.1" "AB761993.1" "AB761992.1"
#> [1386] "AB761991.1" "AB761990.1" "AB761989.1" "AB761988.1" "AB761987.1"
#> [1391] "AB761986.1" "KY459145.1" "KY459130.1" "KY459090.1" "KY459079.1"
#> [1396] "KY459060.1" "KY459033.1" "KY459012.1" "KY458997.1" "MF632278.1"
#> [1401] "MF632277.1" "MF632276.1" "KR902342.1" "LC075899.1" "LC075898.1"
#> [1406] "LC075897.1" "LC075896.1" "LC075895.1" "LC075894.1" "LC075893.1"
#> [1411] "LC075892.1" "LC075891.1" "LC075890.1" "LC075889.1" "LC075888.1"
#> [1416] "LC075887.1" "LC075886.1" "LC075885.1" "AF338376.1" "AF338375.1"
#> [1421] "DQ321670.1" "DQ321668.1" "DQ321667.1" "DQ321663.1" "DQ321662.1"
#> [1426] "DQ098866.1" "DQ098865.1" "DQ098864.1" "DQ098863.1" "DQ098854.1"
#> [1431] "DQ098853.1" "DQ098852.1" "DQ098851.1" "AY429143.1" "AY429142.1"
#> [1436] "JN161068.1" "JN161067.1" "JN161066.1" "JN161065.1" "JN161064.1"
#> [1441] "JN161063.1" "JN161062.1" "HQ005496.1" "HQ005495.1" "HQ005494.1"
#> [1446] "HQ005493.1" "HQ005492.1" "HQ018924.1" "HQ018923.1" "AF181085.1"
#> [1451] "EF658656.1" "EF658655.1" "EF658654.1" "MT483707.1" "MT483705.1"
#> [1456] "MT483704.1" "MT483703.1" "EU179517.1" "EU179516.1" "LC456056.1"
#> [1461] "LC456052.1" "LC456048.1" "LC456044.1" "LC456040.1" "LC456036.1"
#> [1466] "LC456032.1" "LC456028.1" "LC456024.1" "LC456020.1" "LC456016.1"
#> [1471] "LC456012.1" "LC456008.1" "LC456004.1" "LC456000.1" "LC455994.1"
#> [1476] "LC456057.1" "LC456053.1" "LC456049.1" "LC456045.1" "LC456041.1"
#> [1481] "LC456037.1" "LC456033.1" "LC456029.1" "LC456025.1" "LC456021.1"
#> [1486] "LC456017.1" "LC456013.1" "LC456009.1" "LC456005.1" "LC456001.1"
#> [1491] "LC455995.1" "LC456054.1" "LC456050.1" "LC456046.1" "LC456042.1"
#> [1496] "LC456038.1" "LC456034.1" "LC456030.1" "LC456026.1" "LC456022.1"
#> [1501] "LC456018.1" "LC456014.1" "LC456010.1" "LC456006.1" "LC456002.1"
#> [1506] "LC455998.1" "LC455996.1" "LC456055.1" "LC456051.1" "LC456047.1"
#> [1511] "LC456043.1" "LC456039.1" "LC456035.1" "LC456031.1" "LC456027.1"
#> [1516] "LC456023.1" "LC456019.1" "LC456015.1" "LC456011.1" "LC456007.1"
#> [1521] "LC456003.1" "LC455999.1" "LC455997.1" "KY459979.1" "KY459959.1"
#> [1526] "MT488684.1" "MT488683.1" "MT488682.1" "MT488673.1" "MT488668.1"
#> [1531] "MT488667.1" "MT488666.1" "MT488665.1" "MT488664.1" "MT488663.1"
#> [1536] "MT488662.1" "MT488661.1" "MT488660.1" "MT488651.1" "MT488647.1"
#> [1541] "MT488646.1" "MT488645.1" "MT488598.1" "MT488597.1" "MT488596.1"
#> [1546] "MT488595.1" "MT488585.1" "MT488584.1" "MT488583.1" "MT488582.1"
#> [1551] "MT488581.1" "MT488580.1" "MT488579.1" "MT488575.1" "MT488561.1"
#> [1556] "MT488560.1" "MT488559.1" "MT488557.1" "MT488556.1" "MT488555.1"
#> [1561] "MT492195.1" "MT492194.1" "MT492193.1" "MT492192.1" "MT492186.1"
#> [1566] "MT492185.1" "MT492184.1" "MT492183.1" "MT492182.1" "MT492181.1"
#> [1571] "MT492180.1" "MT492176.1" "MT492163.1" "MT492162.1" "MT492161.1"
#> [1576] "MT492160.1" "MT492159.1" "MT492158.1" "MT483795.1" "MT483794.1"
#> [1581] "MT483793.1" "MT483792.1" "MT483782.1" "MT483781.1" "MT483780.1"
#> [1586] "MT483779.1" "MT483778.1" "MT483776.1" "MT483775.1" "MT483772.1"
#> [1591] "MT483768.1" "MT483767.1" "MT483766.1" "MT483764.1" "MT483763.1"
#> [1596] "MT483753.1" "MT483746.1" "MT483745.1" "MT483744.1" "MT483743.1"
#> [1601] "MT483733.1" "MT483732.1" "MT483731.1" "MT483730.1" "MT483729.1"
#> [1606] "MT483728.1" "MT483727.1" "MT483723.1" "MT483709.1" "MT483708.1"
#> [1611] "MT488939.1" "MT488938.1" "MT488937.1" "MT488935.1" "MT488934.1"
#> [1616] "MT488933.1" "MT488931.1" "MT488930.1" "MT488929.1" "MT488928.1"
#> [1621] "MT488919.1" "MT488918.1" "MT488916.1" "MT488915.1" "MT488914.1"
#> [1626] "MT488913.1" "MT488909.1" "MT488896.1" "MT488894.1" "MT488892.1"
#> [1631] "MT488891.1" "MT488890.1" "MT488888.1" "MT488887.1" "MT488886.1"
#> [1636] "MT488885.1" "MT488877.1" "MT488876.1" "MT488875.1" "MT488874.1"
#> [1641] "MT488873.1" "MT488872.1" "MT488871.1" "MT488867.1" "MT488853.1"
#> [1646] "MT488852.1" "MT488851.1" "MT488849.1" "MT488848.1" "MT488847.1"
#> [1651] "MT488845.1" "MT488844.1" "MT488843.1" "MT488842.1" "MT488836.1"
#> [1656] "MT488835.1" "MT488834.1" "MT488833.1" "MT488832.1" "MT488830.1"
#> [1661] "MT488829.1" "MT488826.1" "MT488822.1" "MT488821.1" "MT488820.1"
#> [1666] "MT488819.1" "MT488818.1" "MT488817.1" "MT488806.1" "MT488805.1"
#> [1671] "MT488804.1" "MT488803.1" "MT488794.1" "MT488793.1" "MT488792.1"
#> [1676] "MT488791.1" "MT488790.1" "MT488789.1" "MT488788.1" "MT488784.1"
#> [1681] "MT488770.1" "MT488769.1" "MT488768.1" "MT488766.1" "MT488765.1"
#> [1686] "MT488764.1" "MT488762.1" "MT488761.1" "MT488760.1" "MT488759.1"
#> [1691] "MT488750.1" "MT488749.1" "MT488748.1" "MT488747.1" "MT488746.1"
#> [1696] "MT488745.1" "MT488744.1" "MT488740.1" "MT488726.1" "MT488725.1"
#> [1701] "MT488724.1" "MT488722.1" "MT488721.1" "MT488720.1" "MT489146.1"
#> [1706] "MT489145.1" "MT489144.1" "MT489143.1" "MT489133.1" "MT489132.1"
#> [1711] "MT489131.1" "MT489130.1" "MT489129.1" "MT489128.1" "MT489127.1"
#> [1716] "MT489123.1" "MT489109.1" "MT489108.1" "MT489107.1" "MT489105.1"
#> [1721] "MT489104.1" "MT489103.1" "MT489101.1" "MT489100.1" "MT489099.1"
#> [1726] "MT489098.1" "MT489090.1" "MT489089.1" "MT489088.1" "MT489087.1"
#> [1731] "MT489086.1" "MT489085.1" "MT489084.1" "MT489080.1" "MT489067.1"
#> [1736] "MT489066.1" "MT489065.1" "MT489063.1" "MT489062.1" "MT489061.1"
#> [1741] "MT489059.1" "MT489058.1" "MT489057.1" "MT489056.1" "MT489047.1"
#> [1746] "MT489046.1" "MT489045.1" "MT489044.1" "MT489043.1" "MT489042.1"
#> [1751] "MT489041.1" "MT489037.1" "MT489024.1" "MT489023.1" "MT489022.1"
#> [1756] "MT489020.1" "MT489019.1" "MT489018.1" "MT489016.1" "MT489015.1"
#> [1761] "MT489014.1" "MT489013.1" "MT489005.1" "MT489004.1" "MT489003.1"
#> [1766] "MT489002.1" "MT489001.1" "MT489000.1" "MT488999.1" "MT488995.1"
#> [1771] "MT488982.1" "MT488981.1" "MT488980.1" "MT488978.1" "MT488977.1"
#> [1776] "MT488976.1" "MT488974.1" "MT488973.1" "MT488972.1" "MT488971.1"
#> [1781] "MT488962.1" "MT488961.1" "MT488960.1" "MT488959.1" "MT488958.1"
#> [1786] "MT488957.1" "MT488956.1" "MT488952.1" "HM761930.1" "HM762097.1"
#> [1791] "HM762245.1" "HM762432.1" "HM762516.1" "HM762770.1" "HM762862.1"
#> [1796] "HM763022.1" "HM763266.1" "HM763434.1" "HM763727.1" "HM763997.1"
#> [1801] "HM764105.1" "HM764322.1" "HM764401.1" "HM764649.1" "HM764817.1"
#> [1806] "HM764908.1" "HM765180.1" "U38998.1"   "AY684995.1" "AY684994.1"
#> [1811] "AY684993.1" "AY684992.1" "AY684991.1" "AY646200.1" "AY646198.1"
#> [1816] "DM461196.1" "DM461195.1" "DM461194.1" "DM461193.1" "DM461192.1"
#> [1821] "DM461191.1" "DM367551.1" "DM367548.1" "DM367545.1" "DM367544.1"
#> [1826] "DM367543.1" "DM367542.1" "DM367541.1" "DM367540.1" "DM371283.1"
#> [1831] "DM371280.1" "DM371276.1" "DM371275.1" "DM371274.1" "DM371273.1"
#> [1836] "DM371272.1" "DM371271.1" "DQ334827.1" "HW478258.1" "HW478257.1"
#> [1841] "HW478256.1" "HW478255.1" "DI178960.1" "DI178959.1" "DI178958.1"
#> [1846] "DI178957.1" "DI178956.1" "DI178955.1" "DI178127.1" "DI178124.1"
#> [1851] "DI178120.1" "DI178119.1" "DI178118.1" "DI178117.1" "DI178116.1"
#> [1856] "DI178115.1" "AB239248.1" "AB239247.1" "AB239246.1" "AB239228.1"
#> [1861] "AB239227.1" "AB239226.1" "AB239203.1" "AB239202.1" "AB239201.1"
#> [1866] "AB239200.1" "U59646.1"   "U59645.1"   "U59644.1"   "AF161947.1"
#> [1871] "AF161946.1" "HM756841.1" "HM757384.1" "HM757559.1" "HM757725.1"
#> [1876] "HM758217.1" "HM758302.1" "HM758466.1" "HM758655.1" "HM758935.1"
#> [1881] "HM759106.1" "HM759279.1" "HM759377.1" "HM759647.1" "HM759819.1"
#> [1886] "HM759901.1" "HM760085.1" "HM760261.1" "HM760530.1" "HM760613.1"
#> [1891] "HM760776.1" "HM761014.1" "HM761146.1" "HM761196.1" "HM761512.1"
#> [1896] "HM761773.1" "MT489142.1" "MT489097.1" "MT489055.1" "MT489012.1"
#> [1901] "MT488970.1" "MT488927.1" "MT488884.1" "MT488841.1" "MT488802.1"
#> [1906] "MT488758.1" "MT488674.1" "MT488594.1" "MT492191.1" "MT483742.1"
#> [1911] "KY508415.1" "KY508414.1" "KY508413.1" "KY508412.1" "KY508411.1"
#> [1916] "KY508410.1" "KY508409.1" "KY508408.1" "KY508407.1" "KY508406.1"
#> [1921] "KY508405.1" "KY508404.1" "FJ154795.1" "FJ154794.1" "FJ154793.1"
#> [1926] "DQ098874.1" "DQ098873.1" "DQ098862.1" "DQ098861.1" "AY894656.1"
#> [1931] "AY894655.1" "AY894654.1" "AY894653.1" "AY894652.1" "AY894651.1"
#> [1936] "AY894650.1" "AY894649.1" "AY894648.1" "AY894647.1" "HQ005499.1"
#> [1941] "HQ018926.1" "HM102288.1" "AF004921.1" "HW478260.1" "HW478259.1"
aotus@txids  # taxonomic IDs
#>  [1] "231953"  "1002694" "361674"  "292213"  "57176"   "57175"   "43147"  
#>  [8] "37293"   "280755"  "867331"  "30591"   "120088"  "9505"   
aotus@clstrs # clusters archive
#> Archive of cluster record(s)
#>  - [184] clusters
aotus@sqs    # sequence archive
#> Archive of sequence record(s)
#>  - [1945] sequences
#>  - [13] unique txids
#>  - [597] median sequence length
#>  - [0] median ambiguous nucleotides
aotus@txdct  # taxonomic dictionary
#> Taxonomic dictionary [22] recs, parent [id 9504]
# see all of the available slots
(slotNames(aotus))
#> [1] "cids"    "txids"   "sids"    "txdct"   "sqs"     "clstrs"  "prnt_id"
#> [8] "prnt_nm"
# access different sequences and clusters with [[
(aotus[['0']])              # cluster record 0
#> Cluster Record [id 0]
#>  - [subtree] type
#>  - [AF129795.1] seed sequence
#>  - [173] sequences
#>  - [5] taxa
(aotus[[aotus@sids[[1]]]])  # first sequence record
#> SeqRec [ID: KR528418.1]
# get a summary of the whole object
(summary(aotus))
#>      ID    Type       Seed  Parent N_taxa N_seqs Med_sql       MAD
#> 1     0 subtree AF129795.1    9504      5    173   266.0 0.7090455
#> 2     1 subtree DQ098863.1    9504     10     66  1140.0 0.8966920
#> 3     2 subtree AY659849.1    9504      5     55  1071.0 0.8062280
#> 4     3 subtree DI178118.1    9504      2     52   267.0 0.8666904
#> 5     4 subtree LC075891.1    9504      2     50   844.0 0.8244061
#> 6     5 subtree HM761925.1    9504      9     49   710.0 0.9581328
#> 7     6 subtree HM763429.1    9504      9     49   339.0 0.8417867
#> 8     7 subtree JQ933053.1    9504      9     48   582.5 0.6754071
#> 9     8 subtree HM759641.1    9504      9     48   272.0 0.8482955
#> 10    9 subtree KC761951.1    9504      9     47   649.0 0.9808119
#> 11   10 subtree MT489103.1    9504      8     45  1138.0 0.9992970
#> 12   11 subtree KC762127.1    9504      9     45   797.0 0.9619747
#> 13   12 subtree KC761371.1    9504      9     45   584.0 0.9551948
#> 14   13 subtree KC760228.1    9504      9     43   276.0 0.8748720
#> 15   14 subtree JQ932794.1    9504      4     43   568.0 0.6655075
#> 16   15  direct LC075891.1   30591      1     43   900.0 0.8328878
#> 17   16 subtree EF658652.1    9504      9     41   647.0 0.9793822
#> 18   17 subtree HM760775.1    9504      8     41   680.0 0.9590605
#> 19   18 subtree MT488555.1    9504      8     41   842.0 1.0000000
#> 20   19 subtree   U38998.1    9504      8     40   618.0 0.4020846
#> 21   20 subtree DQ321662.1    9504      8     35   549.0 0.7988127
#> 22   21 subtree JQ932751.1    9504      2     35   796.0 0.7788508
#> 23   22 subtree JN161069.1    9504      1     30  1088.0 0.9985614
#> 24   23  direct LC456000.1   30591      1     30   795.0 0.7748913
#> 25   24 subtree JN161069.1   30591      1     30  1088.0 0.9985614
#> 26   25 subtree AF333711.1    9504      1     28   394.0 0.9228029
#> 27   26 subtree JQ932884.1    9504      4     24   757.5 0.8431763
#> 28   27 subtree LC075890.1    9504      2     24   850.0 0.8423276
#> 29   28 subtree DQ098867.1   30591      2     24  1140.0 0.9663743
#> 30   29 subtree AY900534.1    9504      1     21   291.0 0.8561767
#> 31   30 subtree AF378747.1    9504      1     20   364.0 0.8938575
#> 32   31 subtree AY227055.1 2688256      1     20   269.0 0.9722022
#> 33   32 subtree HM763430.1   30591      2     20   339.0 0.9353022
#> 34   33 subtree HM759643.1   30591      2     20   272.0 0.9726316
#> 35   34 subtree HM761926.1   30591      2     20   711.0 0.9638889
#> 36   35 subtree JQ932877.1    9504      4     19   521.0 0.8932277
#> 37   36 subtree JQ933028.1    9504      2     19   758.0 0.8135355
#> 38   37 subtree JQ932999.1    9504      1     18   633.0 0.6914067
#> 39   38 subtree AF027542.1    9504      1     18   287.0 0.8127515
#> 40   39 subtree AF333712.1    9504      1     18   378.0 0.9713542
#> 41   40  direct LC075890.1   30591      1     18   861.5 0.8424253
#> 42   41 subtree MT483703.1   30591      2     18   618.0 0.9973118
#> 43   42 subtree MT488555.1   30591      2     18   842.0 1.0000000
#> 44   43 subtree MT492158.1   30591      2     18   647.0 0.9997424
#> 45   44 subtree MT483753.1   30591      2     18   905.5 0.9994481
#> 46   45 subtree MT488933.1   30591      2     18   582.5 0.9780015
#> 47   46 subtree MT488817.1   30591      2     18   680.5 0.9938815
#> 48   47 subtree MT489103.1   30591      2     18  1138.0 1.0000000
#> 49   48 subtree MT489061.1   30591      2     18   807.5 0.9813448
#> 50   49 subtree MT489018.1   30591      2     18   649.0 0.9949727
#> 51   50 subtree MT488976.1   30591      2     18   584.0 0.9794101
#> 52   51 subtree JQ932845.1    9504      3     16   587.0 0.6638030
#> 53   52 subtree MT488890.1   30591      2     16   273.0 0.9650350
#> 54   53 subtree MF615377.1    9504      3     14   684.0 0.9140625
#> 55   54  direct LC456005.1   30591      1     14   732.5 0.7900048
#> 56   55 subtree DQ321662.1   30591      2     13   549.0 0.8050398
#> 57   56 subtree AF161946.1    9504      4     12  1059.0 0.9937048
#> 58   57 subtree JQ932800.1    9504      3     12   633.0 0.9001762
#> 59   58 subtree JQ932874.1    9504      1     12   543.5 0.8983452
#> 60   59 subtree KY508404.1    9504      1     12   411.5 0.9586317
#> 61   60 subtree JQ932840.1    9504      1     11   565.0 0.8977661
#> 62   61 subtree AF107750.1    9504      1     11   323.0 0.9789997
#> 63   62 subtree AY900546.1    9504      1     11   290.0 0.6960557
#> 64   63 subtree AY646200.1    9504      4     10   498.0 0.7077341
#> 65   64 subtree JQ932868.1    9504      2     10   561.0 0.8921095
#> 66   65 subtree AY894647.1    9504      1     10  1170.0 0.9992314
#> 67   66 subtree JQ932807.1    9504      1      9   637.0 0.7393674
#> 68   67 subtree AF338375.1    9504      6      8   829.0 0.9398372
#> 69   68 subtree JQ932913.1    9504      2      7   754.0 0.7807910
#> 70   69 subtree HM758650.1    9504      5      7   680.0 0.9595003
#> 71   70 subtree AY900559.1    9504      1      7   285.0 0.7339125
#> 72   71  direct EF658652.1   30591      1      7   629.0 0.9754224
#> 73   72 subtree JQ933020.1    9504      4      6   663.0 0.7837435
#> 74   73 subtree KC761051.1    9504      4      6   473.5 0.6571595
#> 75   74 subtree HM759273.1    9504      5      6   706.5 0.9834499
#> 76   75 subtree AY449069.1    9504      1      6   685.0 1.0000000
#> 77   76 subtree AY684991.1    9504      1      6  1618.0 0.7991497
#> 78   77  direct AY449069.1   30591      1      6   685.0 1.0000000
#> 79   78 subtree DQ989366.1    9504      5      5   428.0 0.9616822
#> 80   79 subtree JQ933005.1    9504      1      5   878.0 0.9155983
#> 81   80 subtree JQ932259.1    9504      1      5   868.0 0.9529944
#> 82   81 subtree JQ932880.1    9504      1      5   556.0 0.8463190
#> 83   82 subtree JQ932869.1    9504      3      5   464.0 0.9661795
#> 84   83 subtree HM756838.1    9504      5      5   549.0 0.9756708
#> 85   84 subtree HM757379.1    9504      5      5   612.0 0.9934641
#> 86   85 subtree HM757554.1    9504      5      5   475.0 0.9233684
#> 87   86 subtree HM757720.1    9504      5      5   898.0 0.9656766
#> 88   87 subtree HM758212.1    9504      5      5   342.0 0.9964912
#> 89   88 subtree HM758297.1    9504      5      5   637.0 0.9885626
#> 90   89 subtree HM758463.1    9504      5      5   590.0 0.9831405
#> 91   90 subtree HM758930.1    9504      5      5   507.0 0.7943343
#> 92   91 subtree HM759101.1    9504      5      5  1077.0 0.9996286
#> 93   92 subtree HM759372.1    9504      5      5   611.0 0.9716150
#> 94   93 subtree HM759729.1    9504      5      5   355.0 0.9684507
#> 95   94 subtree HM759896.1    9504      5      5   580.0 0.9466877
#> 96   95 subtree HM760080.1    9504      5      5   585.0 0.9125413
#> 97   96 subtree HM760256.1    9504      5      5   538.0 0.9951852
#> 98   97 subtree HM760525.1    9504      5      5   554.0 0.9806104
#> 99   98 subtree HM760608.1    9504      5      5   658.0 0.9900452
#> 100  99 subtree HM761193.1    9504      5      5   617.0 0.9765452
#> 101 100 subtree HM761507.1    9504      5      5   558.0 0.9528796
#> 102 101 subtree HM761768.1    9504      5      5   475.0 0.9436326
#> 103 102 subtree HM762092.1    9504      5      5   707.0 0.9786885
#> 104 103 subtree HM762180.1    9504      5      5   761.0 0.9842313
#> 105 104 subtree HM762427.1    9504      5      5   957.0 0.9610152
#> 106 105 subtree HM762765.1    9504      5      5   612.0 0.9719870
#> 107 106 subtree HM763261.1    9504      5      5  1000.0 0.9952000
#> 108 107 subtree HM763727.1    9504      5      5  1245.0 0.9358974
#> 109 108 subtree HM764102.1    9504      5      5   998.0 0.9961924
#> 110 109 subtree HM764317.1    9504      5      5   951.0 0.9501577
#> 111 110 subtree KR902342.1    9504      3      4  1374.0 1.0000000
#> 112 111 subtree AF014508.1    9504      4      4   446.0 0.9983221
#> 113 112 subtree AF014506.1    9504      4      4   312.5 0.9498480
#> 114 113 subtree KC762160.1    9504      4      4   748.0 0.9759680
#> 115 114 subtree KC762072.1    9504      4      4   576.0 0.8039153
#> 116 115 subtree KC762014.1    9504      4      4   663.0 0.9701046
#> 117 116 subtree KC761980.1    9504      4      4   616.5 0.9772000
#> 118 117 subtree KC761918.1    9504      4      4   776.0 0.9932778
#> 119 118 subtree KC761889.1    9504      4      4   522.0 0.9976099
#> 120 119 subtree KC761833.1    9504      4      4   660.0 0.9840504
#> 121 120 subtree KC761782.1    9504      4      4   668.0 0.9589872
#> 122 121 subtree KC761695.1    9504      4      4   735.0 0.8670848
#> 123 122 subtree KC761632.1    9504      4      4   707.0 0.9679219
#> 124 123 subtree KC761601.1    9504      4      4   425.0 0.9982394
#> 125 124 subtree KC761569.1    9504      4      4   744.5 0.9476667
#> 126 125 subtree KC761504.1    9504      4      4   660.0 0.7352784
#> 127 126 subtree KC761470.1    9504      4      4   969.5 0.8648844
#> 128 127 subtree KC761405.1    9504      4      4   459.5 0.8912338
#> 129 128 subtree KC761242.1    9504      4      4   674.0 0.9823529
#> 130 129 subtree KC761177.1    9504      4      4   679.5 0.9255830
#> 131 130 subtree KC761016.1    9504      4      4   445.0 0.9786996
#> 132 131 subtree KC760988.1    9504      4      4   362.0 0.9559229
#> 133 132 subtree JQ932280.1    9504      3      4   554.5 0.7425948
#> 134 133 subtree JQ932662.1    9504      3      4   684.0 0.8431659
#> 135 134 subtree HM762511.1    9504      4      4   521.5 0.9765326
#> 136 135 subtree HM764645.1    9504      4      4   643.5 0.9519231
#> 137 136 subtree AF027546.1    9504      1      4   287.5 0.8907186
#> 138 137 subtree HM764908.1    9504      4      4   490.5 0.9909091
#> 139 138 subtree HM765175.1    9504      4      4   370.0 0.8861985
#> 140 139 subtree AB239200.1    9504      1      4   587.0 0.9995741
#> 141 140  direct DQ098863.1   30591      1      4  1140.0 1.0000000
#> 142 141  direct DQ098851.1   30591      1      4   498.0 1.0000000
#> 143 142 subtree HM758657.1   30591      2      4   679.5 0.9992647
#> 144 143 subtree MW321653.1    9504      2      3  1666.0 0.9988017
#> 145 144 subtree AF014509.1    9504      3      3   343.0 1.0000000
#> 146 145 subtree KC762104.1    9504      3      3   566.0 0.9878261
#> 147 146 subtree KC761277.1    9504      3      3   467.0 0.9775439
#> 148 147 subtree KC761210.1    9504      3      3   386.0 0.8732158
#> 149 148 subtree KC761755.1    9504      3      3   905.0 0.8968421
#> 150 149 subtree KC761438.1    9504      3      3   594.0 0.9977591
#> 151 150 subtree KC761338.1    9504      3      3   538.0 0.9725777
#> 152 151 subtree KC760954.1    9504      3      3   449.0 0.8990349
#> 153 152 subtree KC760866.1    9504      3      3   475.0 0.9845614
#> 154 153 subtree KC760833.1    9504      3      3   527.0 0.9829222
#> 155 154 subtree KC760802.1    9504      3      3   461.0 0.9885305
#> 156 155 subtree KC760770.1    9504      3      3   668.0 0.9736457
#> 157 156 subtree KC760738.1    9504      3      3   518.0 0.9589552
#> 158 157 subtree KC760712.1    9504      3      3   314.0 0.9681529
#> 159 158 subtree KC760681.1    9504      3      3   683.0 0.8234416
#> 160 159 subtree KC760611.1    9504      3      3   531.0 0.9924953
#> 161 160 subtree KC760514.1    9504      3      3   482.0 0.9965565
#> 162 161 subtree KC760417.1    9504      3      3   557.0 0.9976062
#> 163 162 subtree KC760385.1    9504      3      3   420.0 1.0000000
#> 164 163 subtree KC760326.1    9504      3      3   464.0 0.9737589
#> 165 164 subtree KC760263.1    9504      3      3   475.0 0.9971989
#> 166 165 subtree HM761011.1    9504      3      3   764.0 0.9973856
#> 167 166 subtree HM761143.1    9504      3      3   802.0 0.9995844
#> 168 167 subtree HM762858.1    9504      3      3   624.0 0.9941706
#> 169 168 subtree HM763019.1    9504      3      3   431.0 0.9783449
#> 170 169 subtree HM763994.1    9504      3      3   778.0 0.9961489
#> 171 170 subtree HM764396.1    9504      3      3   530.0 1.0000000
#> 172 171 subtree AY900524.1    9504      1      3   366.0 0.7523810
#> 173 172 subtree HM764813.1    9504      3      3   575.0 0.9884660
#> 174 173 subtree AY449065.1    9504      1      3   666.0 0.9980010
#> 175 174 subtree AY449057.1    9504      1      3   667.0 1.0000000
#> 176 175 subtree HM756999.1    9504      3      3   837.0 0.9890325
#> 177 176 subtree AB239246.1    9504      1      3   616.0 0.9989195
#> 178 177 subtree AB239226.1    9504      1      3   614.0 0.9994571
#> 179 178 subtree FJ154793.1    9504      1      3   890.0 0.9992509
#> 180 179  direct AY449065.1   30591      1      3   666.0 0.9980010
#> 181 180  direct AY449057.1   30591      1      3   667.0 1.0000000
#> 182 181  direct MF632276.1   30591      1      3  1059.0 1.0000000
#> 183 182 subtree HM759274.1   30591      2      3   695.0 0.9904489
#> 184 183 subtree AF338376.1   30591      2      3   829.0 1.0000000
#>                            Definition            Feature
#> 1        aotus (0.07), partial (0.07)                  -
#> 2              aotus (0.1), cds (0.1)                  -
#> 3          aotus (0.09), class (0.09)                  -
#> 4   antibody (0.05), construct (0.05) jp (0.3), kr (0.2)
#> 5              dna (0.1), aotus (0.1)                  -
#> 6             aotus (0.1), fbn1 (0.1)                  -
#> 7         aotus (0.09), chrna1 (0.09)                  -
#> 8          aotus (0.1), genomic (0.1)                  -
#> 9             aotus (0.1), gene (0.1)                  -
#> 10         aotus (0.1), genomic (0.1)                  -
#> 11         gene (0.2), complete (0.1)                  -
#> 12         aotus (0.1), genomic (0.1)                  -
#> 13         aotus (0.1), genomic (0.1)                  -
#> 14         aotus (0.1), genomic (0.1)                  -
#> 15         aotus (0.2), genomic (0.2)                  -
#> 16             dna (0.2), aotus (0.1)                  -
#> 17           aotus (0.09), cds (0.09)        barcode (1)
#> 18            aotus (0.1), gene (0.1)                  -
#> 19                atp (0.1), f0 (0.1)                  -
#> 20       gene (0.08), sequence (0.08)                  -
#> 21  gene (0.09), mitochondrial (0.09)                  -
#> 22          aotus (0.1), azarai (0.1)                  -
#> 23          azarai (0.2), aotus (0.1)                  -
#> 24          aotus (0.1), azarai (0.1)                  -
#> 25          azarai (0.2), aotus (0.1)                  -
#> 26           aotus (0.08), cds (0.08)                  -
#> 27         aotus (0.2), genomic (0.2)                  -
#> 28          aotus (0.2), azarai (0.1)                  -
#> 29             aotus (0.1), cds (0.1)                  -
#> 30         aotus (0.08), chain (0.08)                  -
#> 31           aotus (0.08), cds (0.08)                  -
#> 32       antigen (0.08), aotus (0.08)                  -
#> 33        aotus (0.09), azarai (0.09)                  -
#> 34          aotus (0.1), azarai (0.1)                  -
#> 35          aotus (0.1), azarai (0.1)                  -
#> 36         aotus (0.2), genomic (0.2)                  -
#> 37          aotus (0.2), azarai (0.1)                  -
#> 38         aotus (0.2), genomic (0.2)                  -
#> 39         alpha (0.09), aotus (0.09)                  -
#> 40           aotus (0.09), cds (0.09)                  -
#> 41          aotus (0.2), azarai (0.2)                  -
#> 42        aotus (0.08), azarai (0.08)                  -
#> 43                atp (0.1), f0 (0.1)                  -
#> 44        aotus (0.08), azarai (0.08)                  -
#> 45        aotus (0.08), azarai (0.08)                  -
#> 46          aotus (0.1), azarai (0.1)                  -
#> 47          aotus (0.1), azarai (0.1)                  -
#> 48         gene (0.2), complete (0.1)                  -
#> 49          aotus (0.1), azarai (0.1)                  -
#> 50          aotus (0.1), azarai (0.1)                  -
#> 51          aotus (0.1), azarai (0.1)                  -
#> 52         aotus (0.2), genomic (0.2)                  -
#> 53          aotus (0.1), azarai (0.1)                  -
#> 54        allele (0.08), aotus (0.08)                  -
#> 55          aotus (0.2), azarai (0.2)                  -
#> 56          azarai (0.1), gene (0.08)                  -
#> 57            aotus (0.09), cc (0.09)                  -
#> 58         aotus (0.2), genomic (0.2)                  -
#> 59         aotus (0.2), genomic (0.2)                  -
#> 60        sequence (0.2), aotus (0.1)                  -
#> 61         aotus (0.2), genomic (0.2)                  -
#> 62          aotus (0.08), beta (0.08)                  -
#> 63           aotus (0.08), cds (0.08)                  -
#> 64             aotus (0.1), cds (0.1)                  -
#> 65         aotus (0.2), genomic (0.2)                  -
#> 66           aotus (0.1), clone (0.1)                  -
#> 67         aotus (0.2), genomic (0.2)                  -
#> 68             sry (0.2), aotus (0.1)                  -
#> 69         aotus (0.2), genomic (0.2)                  -
#> 70            aotus (0.1), gene (0.1)                  -
#> 71           aotus (0.08), cds (0.08)                  -
#> 72        aotus (0.08), azarai (0.08)                  -
#> 73         aotus (0.2), genomic (0.2)                  -
#> 74         aotus (0.2), genomic (0.2)                  -
#> 75         aotus (0.1), dennd5a (0.1)                  -
#> 76          azarai (0.1), clone (0.1)                  -
#> 77             aotus (0.1), cds (0.1)                  -
#> 78          azarai (0.1), clone (0.1)                  -
#> 79             aotus (0.1), cds (0.1)                  -
#> 80         aotus (0.2), genomic (0.2)                  -
#> 81         aotus (0.2), genomic (0.2)                  -
#> 82         aotus (0.2), genomic (0.2)                  -
#> 83         aotus (0.2), genomic (0.2)                  -
#> 84            aotus (0.1), gene (0.1)                  -
#> 85            aotus (0.1), gene (0.1)                  -
#> 86            aotus (0.1), gene (0.1)                  -
#> 87            aotus (0.1), gene (0.1)                  -
#> 88            aotus (0.1), gene (0.1)                  -
#> 89            aotus (0.1), gene (0.1)                  -
#> 90            aotus (0.1), gene (0.1)                  -
#> 91            aotus (0.1), gene (0.1)                  -
#> 92            aotus (0.1), gene (0.1)                  -
#> 93            aotus (0.1), gene (0.1)                  -
#> 94            aotus (0.1), gene (0.1)                  -
#> 95            aotus (0.1), gene (0.1)                  -
#> 96            aotus (0.1), gene (0.1)                  -
#> 97            aotus (0.1), gene (0.1)                  -
#> 98            aotus (0.1), gene (0.1)                  -
#> 99            aotus (0.1), gene (0.1)                  -
#> 100           aotus (0.1), gene (0.1)                  -
#> 101          aotus (0.1), foxp1 (0.1)                  -
#> 102            aotus (0.1), fes (0.1)                  -
#> 103        aotus (0.1), fam123b (0.1)                  -
#> 104           aotus (0.1), erc2 (0.1)                  -
#> 105           aotus (0.1), edg1 (0.1)                  -
#> 106          aotus (0.1), dctn2 (0.1)                  -
#> 107           aotus (0.1), cnr1 (0.1)                  -
#> 108          aotus (0.1), brca2 (0.1)                  -
#> 109           aotus (0.1), bche (0.1)                  -
#> 110          aotus (0.1), axin1 (0.1)                  -
#> 111            cd4 (0.2), aotus (0.1)                  -
#> 112          alpha (0.1), aotus (0.1)                  -
#> 113            aotus (0.1), cds (0.1)                  -
#> 114        aotus (0.2), genomic (0.2)                  -
#> 115        aotus (0.2), genomic (0.2)                  -
#> 116        aotus (0.2), genomic (0.2)                  -
#> 117        aotus (0.2), genomic (0.2)                  -
#> 118        aotus (0.2), genomic (0.2)                  -
#> 119        aotus (0.2), genomic (0.2)                  -
#> 120        aotus (0.2), genomic (0.2)                  -
#> 121        aotus (0.2), genomic (0.2)                  -
#> 122        aotus (0.2), genomic (0.2)                  -
#> 123        aotus (0.2), genomic (0.2)                  -
#> 124        aotus (0.2), genomic (0.2)                  -
#> 125        aotus (0.2), genomic (0.2)                  -
#> 126        aotus (0.2), genomic (0.2)                  -
#> 127        aotus (0.2), genomic (0.2)                  -
#> 128        aotus (0.2), genomic (0.2)                  -
#> 129        aotus (0.2), genomic (0.2)                  -
#> 130        aotus (0.2), genomic (0.2)                  -
#> 131        aotus (0.2), genomic (0.2)                  -
#> 132        aotus (0.2), genomic (0.2)                  -
#> 133        aotus (0.2), genomic (0.2)                  -
#> 134        aotus (0.2), genomic (0.2)                  -
#> 135          aotus (0.1), dmrt1 (0.1)                  -
#> 136            aotus (0.1), app (0.1)                  -
#> 137        alpha (0.09), aotus (0.09)                  -
#> 138           aff2 (0.1), aotus (0.1)                  -
#> 139         adora3 (0.1), aotus (0.1)                  -
#> 140            aotus (0.1), dna (0.1)                  -
#> 141         aotus (0.1), azarai (0.1)                  -
#> 142         aotus (0.1), azarai (0.1)                  -
#> 143         aotus (0.1), azarai (0.1)                  -
#> 144            aotus (0.1), cds (0.1)                  -
#> 145            aotus (0.1), cds (0.1)                  -
#> 146        aotus (0.2), genomic (0.2)                  -
#> 147        aotus (0.2), genomic (0.2)                  -
#> 148        aotus (0.2), genomic (0.2)                  -
#> 149        aotus (0.2), genomic (0.2)                  -
#> 150        aotus (0.2), genomic (0.2)                  -
#> 151        aotus (0.2), genomic (0.2)                  -
#> 152        aotus (0.2), genomic (0.2)                  -
#> 153        aotus (0.2), genomic (0.2)                  -
#> 154        aotus (0.2), genomic (0.2)                  -
#> 155        aotus (0.2), genomic (0.2)                  -
#> 156        aotus (0.2), genomic (0.2)                  -
#> 157        aotus (0.2), genomic (0.2)                  -
#> 158        aotus (0.2), genomic (0.2)                  -
#> 159        aotus (0.2), genomic (0.2)                  -
#> 160        aotus (0.2), genomic (0.2)                  -
#> 161        aotus (0.2), genomic (0.2)                  -
#> 162        aotus (0.2), genomic (0.2)                  -
#> 163        aotus (0.2), genomic (0.2)                  -
#> 164        aotus (0.2), genomic (0.2)                  -
#> 165        aotus (0.2), genomic (0.2)                  -
#> 166           aotus (0.1), gene (0.1)                  -
#> 167           aotus (0.1), gene (0.1)                  -
#> 168          aotus (0.1), dach1 (0.1)                  -
#> 169           aotus (0.1), crem (0.1)                  -
#> 170           aotus (0.1), bcor (0.1)                  -
#> 171          aotus (0.1), atxn7 (0.1)                  -
#> 172          aotus (0.08), cds (0.08)                  -
#> 173          aff22 (0.1), aotus (0.1)                  -
#> 174         autos (0.1), azarai (0.1)                  -
#> 175         autos (0.1), azarai (0.1)                  -
#> 176         aotus (0.1), azarai (0.1)                  -
#> 177            aotus (0.1), dna (0.1)                  -
#> 178            aotus (0.1), dna (0.1)                  -
#> 179          aotus (0.08), cds (0.08)                  -
#> 180         autos (0.1), azarai (0.1)                  -
#> 181         autos (0.1), azarai (0.1)                  -
#> 182         aotus (0.1), azarai (0.1)                  -
#> 183         aotus (0.1), azarai (0.1)                  -
#> 184            sry (0.2), aotus (0.1)                  -
# the above generates a data.frame with information on each cluster:
# ID - unique id in the object
# Type - cluster type
# Seed - most connected sequence
# Parent - MRCA of all represented taxa
# N_taxa - number of NCBI recognised taxa
# N_seqs - number of sequences
# Med_sql - median sequence length
# MAD - Maximum alignment density, values close to 1 indicate all sequences in
#  the cluster have a similar length.
# Definition - most common words (and frequency) in sequence definitions
# Feature - most common feature name (and frequency)